refactor(cli): split cli.py into per-stage submodules (#221)
Closes #221.

The 2,169-line `src/nhf_spatial_targets/cli.py` is split along the
four cyclopts sub-app boundaries it already had internally:

- `cli/__init__.py`  (126) — `app = App(...)`, sub-app wiring, the
  `launcher` meta.default, and re-exports of `_dispatch`,
  `setup_logging`, `materialize_credentials_cmd`, and every
  `aggregate_*` callable so existing tests that
  `patch("nhf_spatial_targets.cli.aggregate_<src>", ...)` continue to
  intercept after the split. `main = app.meta` is preserved for the
  `pyproject.toml` `[project.scripts]` entry.
- `cli/_params.py`   ( 40) — shared `Annotated[..., Parameter(...)]`
  aliases for the agg sub-app (`_AGG_BATCH_SIZE_PARAM`,
  `_AGG_WORKER_INDEX_PARAM`, `_AGG_N_WORKERS_PARAM`).
- `cli/fetch.py`     (916) — `fetch_app` and its 15 commands.
- `cli/agg.py`       (665) — `agg_app` and its 16 commands plus the
  `_run_tier_agg` and `_resolve_agg_config` helpers. Per-source
  `aggregate_*` callables are resolved via `_resolve_agg_fn(name)`,
  which `getattr`s them off `nhf_spatial_targets.cli` at
  command-invocation time so the patch indirection holds.
- `cli/catalog.py`   ( 25) — `catalog_app` plus its two inspection
  commands.
- `cli/run.py`       (565) — root-level commands: `run`, `rechunk`,
  `reconcile-manifest`, `upgrade-config`, `init`,
  `materialize-credentials`, `validate`. The `_dispatch` helper for
  `run` is co-located here and re-exported from `__init__.py`.

No behavior change. Every cyclopts command name, alias, default, and
help text is byte-identical. Verified with `pixi run nhf-targets
{--help, fetch --help, agg --help, catalog --help}` and 107 CLI-
touching tests (`test_cli{,_agg,_materialize_credentials}.py` +
`test_aggregate_daymet.py` + `test_upgrade_config.py` +
`test_reitz2017.py`), all pass.

`fetch.py` (916 LOC) and `agg.py` (665 LOC) are notably above the
~500-LOC target in the issue. The overshoot is intentional: per-
command boilerplate (print-fetching / try-except / print-success +
JSON) varies just enough across sources that consolidating it into a
helper risks subtle behavior change. Issue #221 specifies "pure file
split" — further mechanical deduping is a separate ticket.
