--- src/nhf_spatial_targets/cli.py
@@ -1,2169 +1,0 @@
-"""nhf-targets command-line interface."""
-
-from __future__ import annotations
-
-import logging
-import sys
-from collections.abc import Callable
-from importlib.metadata import version as _pkg_version
-from pathlib import Path
-from typing import TYPE_CHECKING, Annotated
-
-if TYPE_CHECKING:
-    from nhf_spatial_targets.workspace import Project
-
-
-import yaml
-from cyclopts import App, Parameter
-
-from nhf_spatial_targets._logging import setup_logging
-from nhf_spatial_targets.aggregate.daymet import aggregate_daymet
-from nhf_spatial_targets.aggregate.era5_land import (
-    aggregate_era5_land,
-    aggregate_era5_land_sd,
-)
-from nhf_spatial_targets.aggregate.gldas import aggregate_gldas
-from nhf_spatial_targets.aggregate.margulis_wus_sr import aggregate_margulis_wus_sr
-from nhf_spatial_targets.aggregate.merra2 import aggregate_merra2
-from nhf_spatial_targets.aggregate.mod10c1 import aggregate_mod10c1
-from nhf_spatial_targets.aggregate.mod16a2 import aggregate_mod16a2
-from nhf_spatial_targets.aggregate.mwbm_climgrid import aggregate_mwbm_climgrid
-from nhf_spatial_targets.aggregate.ncep_ncar import aggregate_ncep_ncar
-from nhf_spatial_targets.aggregate.nldas_mosaic import aggregate_nldas_mosaic
-from nhf_spatial_targets.aggregate.nldas_noah import aggregate_nldas_noah
-from nhf_spatial_targets.aggregate.reitz2017 import aggregate_reitz2017
-from nhf_spatial_targets.aggregate.snodas import aggregate_snodas
-from nhf_spatial_targets.aggregate.watergap22d import aggregate_watergap22d
-
-_logger = logging.getLogger(__name__)
-
-# Shared agg-command parameter metadata. Each command annotates its
-# matching parameter as ``Annotated[<type>, _AGG_*_PARAM] = <default>``.
-# Cyclopts reads name + help from the Parameter; the default lives on
-# the function signature so the type checker still sees a concrete
-# default (a single source of truth for both behavior + docs).
-_AGG_BATCH_SIZE_PARAM = Parameter(
-    name="--batch-size",
-    help=(
-        "Target HRUs per spatial batch. Overrides ``fabric.batch_size`` "
-        "from ``config.yml`` (default 500). Changing this value "
-        "invalidates cached weight CSVs via the SHA-256 batch-HRU "
-        "fingerprint."
-    ),
-)
-_AGG_WORKER_INDEX_PARAM = Parameter(
-    name=["--worker-index", "-w"],
-    help=(
-        "SLURM-array round-robin worker index in [0, --n-workers). "
-        "Default 0 (single worker). Each worker processes a disjoint "
-        "round-robin slice of the source's years; see "
-        "docs/sources/ on the SLURM array recipe (issue #156)."
-    ),
-)
-_AGG_N_WORKERS_PARAM = Parameter(
-    name=["--n-workers", "-n"],
-    help=(
-        "Total SLURM-array workers for round-robin year sharding. "
-        "Default 1 (serial). Must match the SLURM ``--array`` upper "
-        "bound + 1. Each worker writes its assigned per-year NCs and "
-        "merges its slice into manifest.json under a flock."
-    ),
-)
-
-app = App(
-    name="nhf-targets",
-    help="nhf-spatial-targets: build NHM calibration target datasets.",
-    version=_pkg_version("nhf-spatial-targets"),
-)
-fetch_app = App(name="fetch", help="Download source datasets into a project datastore.")
-agg_app = App(name="agg", help="Aggregate source datasets to HRU fabric polygons.")
-catalog_app = App(name="catalog", help="Inspect the data source catalog.")
-app.command(fetch_app)
-app.command(agg_app)
-app.command(catalog_app)
-
-
-@app.meta.default
-def launcher(
-    *tokens: Annotated[str, Parameter(show=False, allow_leading_hyphen=True)],
-    verbose: Annotated[bool, Parameter(name=["--verbose", "-v"])] = False,
-):
-    """Global options for nhf-targets."""
-    setup_logging(verbose)
-    app(tokens)  # dispatch remaining tokens to the root app
-
-
-@app.command
-def run(
-    workdir: Annotated[
-        Path,
-        Parameter(
-            name=["--project-dir", "-d"],
-            help="Project created by 'nhf-targets init'.",
-        ),
-    ],
-    target: Annotated[
-        str | None,
-        Parameter(
-            name=["--target", "-t"],
-            help="Run a single target (default: all enabled).",
-        ),
-    ] = None,
-):
-    """Run the calibration target pipeline."""
-    from nhf_spatial_targets.workspace import load as load_project
-
-    if not workdir.exists():
-        print(f"Error: Project not found: {workdir}", file=sys.stderr)
-        sys.exit(2)
-    if not (workdir / "fabric.json").exists():
-        print(
-            f"Error: fabric.json not found in {workdir}. "
-            "Run 'nhf-targets validate' first.",
-            file=sys.stderr,
-        )
-        sys.exit(2)
-
-    try:
-        project = load_project(workdir)
-    except (FileNotFoundError, ValueError) as exc:
-        print(f"Error: {exc}", file=sys.stderr)
-        sys.exit(1)
-
-    targets_cfg = project.config.get("targets", {})
-    to_run = (
-        [target]
-        if target
-        else [k for k, v in targets_cfg.items() if v.get("enabled", False)]
-    )
-
-    for name in to_run:
-        if name not in targets_cfg:
-            print(f"Error: Unknown target: {name}", file=sys.stderr)
-            sys.exit(1)
-        print(f"Building target: {name}")
-        try:
-            _dispatch(name, project)
-        except NotImplementedError as exc:
-            print(
-                f"WARNING: target '{name}' not yet implemented; skipping ({exc})",
-                file=sys.stderr,
-            )
-            continue
-        except Exception as exc:
-            _logger.exception("Error building target '%s'", name)
-            print(f"Error building target '{name}': {exc}", file=sys.stderr)
-            sys.exit(1)
-
-
-def _dispatch(
-    name: str,
-    project: "Project",
-) -> None:
-    """Dispatch to the appropriate target builder module.
-
-    All builders share the ``build(project: Project) -> None`` signature
-    of :func:`targets.run.build`; per-target config is read from
-    ``project.target(name)``.
-    """
-    from nhf_spatial_targets.targets import aet, rch, run, sca, som, swe
-
-    builders = {
-        "runoff": run.build,
-        "aet": aet.build,
-        "recharge": rch.build,
-        "soil_moisture": som.build,
-        "snow_covered_area": sca.build,
-        "snow_water_equivalent": swe.build,
-    }
-    if name not in builders:
-        print(f"Error: No builder registered for target: {name}", file=sys.stderr)
-        sys.exit(1)
-    builders[name](project)
-
-
-@app.command
-def rechunk(
-    workdir: Annotated[
-        Path,
-        Parameter(
-            name=["--project-dir", "-d"],
-            help="Project created by 'nhf-targets init'.",
-        ),
-    ],
-    layer: Annotated[
-        str | None,
-        Parameter(
-            name=["--layer"],
-            help="Restrict to 'aggregated' or 'target' (default: both).",
-        ),
-    ] = None,
-    source: Annotated[
-        str | None,
-        Parameter(
-            name=["--source"],
-            help="Restrict the aggregated layer to one source key.",
-        ),
-    ] = None,
-    dry_run: Annotated[
-        bool,
-        Parameter(
-            name=["--dry-run"],
-            help="Report what would be rechunked without writing.",
-        ),
-    ] = False,
-):
-    """Backfill existing aggregated/target NCs to the chunked+compressed layout.
-
-    Rewrites contiguous/uncompressed NetCDFs (built before #165) in place to the
-    canonical io_nc layout: value-preserving, atomic, and idempotent. Does not
-    touch the shared datastore's consolidated NCs, nor the daymet/ssebop
-    aggregated outputs (left as-is by #165 ST3).
-    """
-    from nhf_spatial_targets.rechunk import rechunk_project
-    from nhf_spatial_targets.workspace import load as load_project
-
-    if not workdir.exists():
-        print(f"Error: Project not found: {workdir}", file=sys.stderr)
-        sys.exit(2)
-    if not (workdir / "fabric.json").exists():
-        print(
-            f"Error: fabric.json not found in {workdir}. "
-            "Run 'nhf-targets validate' first.",
-            file=sys.stderr,
-        )
-        sys.exit(2)
-    if layer is not None and layer not in ("aggregated", "target"):
-        print(
-            f"Error: invalid --layer {layer!r}; expected 'aggregated' or 'target'.",
-            file=sys.stderr,
-        )
-        sys.exit(2)
-
-    try:
-        project = load_project(workdir)
-    except (FileNotFoundError, ValueError) as exc:
-        print(f"Error: {exc}", file=sys.stderr)
-        sys.exit(1)
-
-    # A mistyped --source would otherwise be a silent no-op; fail loudly.
-    if source is not None and not (project.aggregated_dir() / source).is_dir():
-        print(
-            f"Error: no aggregated source directory '{source}' under "
-            f"{project.aggregated_dir()}.",
-            file=sys.stderr,
-        )
-        sys.exit(2)
-
-    results = rechunk_project(project, layer=layer, source=source, dry_run=dry_run)
-
-    rechunked = [r for r in results if r["status"] == "rechunked"]
-    skipped = [r for r in results if r["status"].startswith("skipped")]
-    planned = [r for r in results if r["status"] == "would-rechunk"]
-    failed = [r for r in results if r["status"] == "failed"]
-    total_before = sum(r["size_before"] for r in results)
-    total_after = sum(r["size_after"] for r in results if r["size_after"] is not None)
-
-    if dry_run:
-        print(
-            f"[dry-run] {len(planned)} file(s) would be rechunked, "
-            f"{len(skipped)} skipped "
-            f"({total_before / 1e9:.2f} GB candidate)."
-        )
-    else:
-        saved = total_before - total_after if rechunked else 0
-        reclaimed_from = sum(r["size_before"] for r in rechunked)
-        pct = 100.0 * saved / reclaimed_from if reclaimed_from else 0.0
-        print(
-            f"Rechunked {len(rechunked)} file(s), skipped {len(skipped)}. "
-            f"Reclaimed {saved / 1e9:.2f} GB "
-            f"({pct:.0f}% of the rewritten files)."
-        )
-
-    if failed:
-        print(f"Error: {len(failed)} file(s) failed to rechunk:", file=sys.stderr)
-        for r in failed:
-            print(f"  {r['path'].name}: {r.get('error')}", file=sys.stderr)
-        sys.exit(1)
-
-
-@app.command(name="reconcile-manifest")
-def reconcile_manifest_cmd(
-    workdir: Annotated[
-        Path,
-        Parameter(
-            name=["--project-dir", "-d"],
-            help="Project created by 'nhf-targets init'.",
-        ),
-    ],
-    source: Annotated[
-        list[str] | None,
-        Parameter(
-            name=["--source"],
-            help="Catalog source key to reconcile (repeatable). Default: all.",
-        ),
-    ] = None,
-    dry_run: Annotated[
-        bool,
-        Parameter(name=["--dry-run"], help="Report what would change; write nothing."),
-    ] = False,
-    checksum: Annotated[
-        bool,
-        Parameter(name=["--checksum"], help="Compute sha256 for each record (slow)."),
-    ] = False,
-):
-    """Backfill manifest.json from consolidated NCs already in the datastore.
-
-    Use after creating a new project against a datastore that another project
-    already populated. Adds 'provenance: reconciled' file records for sources
-    found on disk but missing from this project's manifest; never overwrites
-    existing records. See docs/architecture/reconcile-manifest.md.
-    """
-    from rich.console import Console
-    from rich.table import Table
-
-    from nhf_spatial_targets import workspace
-    from nhf_spatial_targets.reconcile import reconcile_manifest
-
-    console = Console()
-    if not workdir.exists():
-        print(f"Error: Project not found: {workdir}", file=sys.stderr)
-        sys.exit(2)
-
-    try:
-        project = workspace.load(workdir)
-        results = reconcile_manifest(
-            project, sources=source, dry_run=dry_run, checksum=checksum
-        )
-    except (FileNotFoundError, ValueError, KeyError, OSError) as e:
-        print(f"reconcile-manifest failed: {e}", file=sys.stderr)
-        sys.exit(1)
-
-    title = (
-        "reconcile-manifest (dry run — no changes written)"
-        if dry_run
-        else "reconcile-manifest"
-    )
-    table = Table(title=title)
-    table.add_column("source", style="bold")
-    table.add_column("status")
-    table.add_column("on disk", justify="right")
-    table.add_column("already recorded", justify="right")
-    table.add_column("added", justify="right")
-    for r in results:
-        table.add_row(
-            r.source_key,
-            r.status,
-            str(r.on_disk),
-            str(r.already_recorded),
-            str(r.added),
-        )
-    console.print(table)
-
-    total_added = sum(r.added for r in results)
-    verb = "would add" if dry_run else "added"
-    console.print(
-        f"[bold green]{verb} {total_added} reconciled record(s).[/bold green]"
-    )
-
-
-@app.command(name="upgrade-config")
-def upgrade_config_cmd(
-    workdir: Annotated[
-        Path,
-        Parameter(
-            name=["--project-dir", "-d"],
-            help="Project created by 'nhf-targets init'.",
-        ),
-    ],
-):
-    """Report optional-config features missing from the project's config.yml.
-
-    Existing projects don't pick up new optional features (e.g. fabric.token,
-    representative_points) added to the init template, because they were
-    created before those features existed. This command compares the project's
-    config.yml against the registry in upgrade_config.OPTIONAL_CONFIG_FEATURES
-    and prints the literal commented block to paste for each missing feature.
-
-    Report-only: never mutates the operator's config.yml. Exits 0 if in sync,
-    1 on drift (so scripted heartbeats can detect it).
-    """
-    from rich.console import Console
-    from rich.table import Table
-
-    from nhf_spatial_targets.upgrade_config import check_drift
-
-    console = Console()
-    if not workdir.exists():
-        print(f"Error: Project not found: {workdir}", file=sys.stderr)
-        sys.exit(2)
-    try:
-        missing = check_drift(workdir)
-    except FileNotFoundError as e:
-        print(f"upgrade-config failed: {e}", file=sys.stderr)
-        sys.exit(1)
-
-    if not missing:
-        console.print(
-            "[bold green]Project config is in sync with the latest "
-            "optional-feature stubs in the init template.[/bold green]"
-        )
-        return
-
-    table = Table(title="Optional features missing from this project's config.yml")
-    table.add_column("feature", style="bold")
-    table.add_column("added")
-    table.add_column("why")
-    for feat in missing:
-        table.add_row(feat.name, feat.added, feat.why)
-    console.print(table)
-
-    console.print(
-        "\nPaste each block below into config.yml (or see "
-        "src/nhf_spatial_targets/init_run.py:_CONFIG_TEMPLATE for context). "
-        "This command never edits your file.\n"
-    )
-    for feat in missing:
-        console.print(f"[bold]# --- {feat.name} ---[/bold]")
-        console.print(feat.block)
-    sys.exit(1)
-
-
-@app.command
-def init(
-    workdir: Annotated[
-        Path,
-        Parameter(
-            name=["--project-dir", "-d"],
-            help="Directory to create as the new project.",
-        ),
-    ],
-):
-    """Initialise a new project with a config template.
-
-    Creates a directory skeleton with config.yml and .credentials.yml.
-    Edit those files, then run 'nhf-targets validate --project-dir <dir>'.
-    """
-    from nhf_spatial_targets.init_run import init_project
-    from rich.console import Console
-    from rich.panel import Panel
-    from rich.text import Text
-
-    console = Console()
-
-    try:
-        result = init_project(workdir)
-    except FileExistsError as e:
-        print(f"Error: {e}", file=sys.stderr)
-        sys.exit(1)
-
-    msg = Text()
-    msg.append("Project created:\n", style="bold green")
-    msg.append(f"  {result}\n\n")
-    msg.append("Next steps:\n", style="bold")
-    msg.append(f"  1. Edit   {result / 'config.yml'}\n")
-    msg.append(f"  2. Fill   {result / '.credentials.yml'}\n")
-    msg.append(f"  3. Run    nhf-targets validate --project-dir {result}\n")
-    console.print(Panel(msg, title="nhf-targets init", border_style="green"))
-
-
-@app.command(name="materialize-credentials")
-def materialize_credentials_cmd(
-    workdir: Annotated[
-        Path,
-        Parameter(
-            name=["--project-dir", "-d"],
-            help="Project directory containing .credentials.yml.",
-        ),
-    ],
-):
-    """Copy credentials from .credentials.yml into ~/.cdsapirc and ~/.netrc.
-
-    Reads the 'cds' and 'nasa_earthdata' sections from the project's
-    .credentials.yml and writes the corresponding dotfiles consumed by
-    cdsapi and earthaccess at runtime.
-
-    Both files are written atomically and set to mode 0600.  Run this command
-    after editing or rotating .credentials.yml.
-
-    Each section (cds, nasa_earthdata) is processed independently — one
-    section failing does not prevent the other from being written.  The
-    command exits non-zero if any section fails.
-
-    Exit codes:
-      0 — all sections written successfully
-      1 — incomplete credentials (ValueError) — user action required
-      2 — project directory not found
-      3 — write failure (OSError) — system action required
-    """
-    from rich.console import Console
-    from rich.table import Table
-
-    from nhf_spatial_targets.credentials import (
-        materialize_cdsapirc,
-        materialize_netrc_earthdata,
-    )
-
-    console = Console()
-
-    cred_path = workdir / ".credentials.yml"
-    if not workdir.exists():
-        print(f"Error: Project not found: {workdir}", file=sys.stderr)
-        sys.exit(2)
-    if not cred_path.exists():
-        print(
-            f"Error: .credentials.yml not found in {workdir}. "
-            "Run 'nhf-targets init --project-dir <dir>' to create a template.",
-            file=sys.stderr,
-        )
-        sys.exit(1)
-
-    try:
-        raw = yaml.safe_load(cred_path.read_text())
-    except yaml.YAMLError as exc:
-        print(f"Error: Cannot parse {cred_path}: {exc}", file=sys.stderr)
-        sys.exit(1)
-    if raw is None:
-        print(
-            f"Error: {cred_path} is empty — did you save your edits?",
-            file=sys.stderr,
-        )
-        sys.exit(1)
-    creds = raw if isinstance(raw, dict) else {}
-
-    table = Table(title="Credential materialisation", show_header=True)
-    table.add_column("Section", style="bold")
-    table.add_column("Target file")
-    table.add_column("Status")
-
-    errors: list[str] = []
-
-    # --- CDS ---
-    try:
-        cds_path = materialize_cdsapirc(creds)
-        table.add_row("cds", str(cds_path), "[green]written[/green]")
-    except ValueError as exc:
-        msg = f"{cred_path}: {exc}"
-        table.add_row("cds", "~/.cdsapirc", f"[yellow]skipped[/yellow]: {exc}")
-        errors.append(("user", msg))
-    except OSError as exc:
-        msg = f"{cred_path}: {exc}"
-        table.add_row("cds", "~/.cdsapirc", f"[red]error[/red]: {exc}")
-        errors.append(("system", msg))
-
-    # --- NASA Earthdata ---
-    try:
-        netrc_path = materialize_netrc_earthdata(creds)
-        table.add_row("nasa_earthdata", str(netrc_path), "[green]written[/green]")
-    except ValueError as exc:
-        msg = f"{cred_path}: {exc}"
-        table.add_row("nasa_earthdata", "~/.netrc", f"[yellow]skipped[/yellow]: {exc}")
-        errors.append(("user", msg))
-    except OSError as exc:
-        msg = f"{cred_path}: {exc}"
-        table.add_row("nasa_earthdata", "~/.netrc", f"[red]error[/red]: {exc}")
-        errors.append(("system", msg))
-
-    console.print(table)
-
-    if errors:
-        has_system_error = any(kind == "system" for kind, _ in errors)
-        has_user_error = any(kind == "user" for kind, _ in errors)
-        if has_user_error:
-            console.print(
-                "\n[yellow]One or more sections were skipped due to missing or "
-                "incomplete credentials.  Fill in .credentials.yml and re-run.[/yellow]"
-            )
-        if has_system_error:
-            console.print(
-                "\n[red]One or more sections failed due to a system error "
-                "(e.g. filesystem permissions).  See the table above for details.[/red]"
-            )
-        sys.exit(3 if has_system_error else 1)
-
-    console.print(
-        "\n[bold green]All credentials materialised successfully.[/bold green]"
-    )
-
-
-@app.command
-def validate(
-    workdir: Annotated[
-        Path,
-        Parameter(
-            name=["--project-dir", "-d"],
-            help="Project directory to validate.",
-        ),
-    ],
-):
-    """Validate a project: check config, fabric, credentials, and catalog.
-
-    On success, writes fabric.json and manifest.json into the project.
-    """
-    from rich.console import Console
-
-    from nhf_spatial_targets.validate import validate_workspace
-
-    console = Console()
-
-    if not workdir.exists():
-        print(f"Error: Project not found: {workdir}", file=sys.stderr)
-        sys.exit(2)
-
-    try:
-        validate_workspace(workdir)
-    except (FileNotFoundError, ValueError, RuntimeError, OSError) as e:
-        print(f"Validation failed: {e}", file=sys.stderr)
-        sys.exit(1)
-    except Exception as exc:
-        _logger.exception("Unexpected error during validation")
-        print(
-            f"Unexpected validation error ({type(exc).__name__}): {exc}",
-            file=sys.stderr,
-        )
-        sys.exit(1)
-
-    console.print(f"[bold green]Project validated successfully:[/bold green] {workdir}")
-
-
-@fetch_app.command(name="all")
-def fetch_all_cmd(
-    workdir: Annotated[
-        Path,
-        Parameter(
-            name=["--project-dir"],
-            help="Project created by 'nhf-targets init'.",
-        ),
-    ],
-    period: Annotated[
-        str,
-        Parameter(name=["--period", "-p"], help="Temporal range as 'YYYY/YYYY'."),
-    ],
-):
-    """Download all source datasets into the project datastore.
-
-    Iterates through every registered fetch module in sequence.
-    Stops on the first failure.
-    """
-    from rich.console import Console
-
-    if not workdir.exists():
-        print(f"Error: Project not found: {workdir}", file=sys.stderr)
-        sys.exit(2)
-
-    console = Console()
-
-    # Import all fetch functions
-    import nhf_spatial_targets.catalog as _catalog
-    from nhf_spatial_targets.fetch._period import clamp_period
-    from nhf_spatial_targets.fetch.daymet import fetch_daymet
-    from nhf_spatial_targets.fetch.era5_land import fetch_era5_land
-    from nhf_spatial_targets.fetch.gldas import fetch_gldas
-    from nhf_spatial_targets.fetch.margulis_wus_sr import fetch_margulis_wus_sr
-    from nhf_spatial_targets.fetch.merra2 import fetch_merra2
-    from nhf_spatial_targets.fetch.modis import fetch_mod10c1, fetch_mod16a2
-    from nhf_spatial_targets.fetch.mwbm_climgrid import fetch_mwbm_climgrid
-    from nhf_spatial_targets.fetch.ncep_ncar import fetch_ncep_ncar
-    from nhf_spatial_targets.fetch.nldas import fetch_nldas_mosaic, fetch_nldas_noah
-    from nhf_spatial_targets.fetch.pangaea import fetch_watergap22d
-    from nhf_spatial_targets.fetch.reitz2017 import fetch_reitz2017
-    from nhf_spatial_targets.fetch.snodas import fetch_snodas
-
-    # (display name, catalog source key, fetch function)
-    sources = [
-        ("era5-land", "era5_land", fetch_era5_land),
-        ("gldas", "gldas_noah_v21_monthly", fetch_gldas),
-        ("merra2", "merra2", fetch_merra2),
-        ("nldas-mosaic", "nldas_mosaic", fetch_nldas_mosaic),
-        ("nldas-noah", "nldas_noah", fetch_nldas_noah),
-        ("ncep-ncar", "ncep_ncar", fetch_ncep_ncar),
-        ("mod16a2", "mod16a2_v061", fetch_mod16a2),
-        ("mod10c1", "mod10c1_v061", fetch_mod10c1),
-        ("watergap22d", "watergap22d", fetch_watergap22d),
-        ("reitz2017", "reitz2017", fetch_reitz2017),
-        ("mwbm-climgrid", "mwbm_climgrid", fetch_mwbm_climgrid),
-        # SWE category (issue #99) — fetch scaffolding only; aggregate +
-        # target wiring deferred to per-source follow-up issues.
-        ("daymet", "daymet", fetch_daymet),
-        ("snodas", "snodas", fetch_snodas),
-        ("margulis-wus-sr", "margulis_wus_sr", fetch_margulis_wus_sr),
-    ]
-
-    results = {}
-    for name, source_key, fetch_fn in sources:
-        console.print(f"\n[bold]{'─' * 60}[/bold]")
-
-        # Clamp requested period to the source's available range
-        meta = _catalog.source(source_key)
-        available = meta.get("period", "")
-        clamped = clamp_period(period, available) if available else period
-
-        if clamped is None:
-            console.print(
-                f"[yellow]{name}: skipped (no overlap between "
-                f"requested {period} and available {available})[/yellow]"
-            )
-            continue
-
-        if clamped != period:
-            console.print(
-                f"[bold]Fetching {name} for period {clamped} "
-                f"(clamped from {period} to available {available})...[/bold]"
-            )
-        else:
-            console.print(f"[bold]Fetching {name} for period {period}...[/bold]")
-
-        try:
-            result = fetch_fn(workdir=workdir, period=clamped)
-            results[name] = result
-            console.print(f"[green]{name}: downloaded to datastore[/green]")
-        except FileNotFoundError as exc:
-            # Manual-placement / operator-staged sources can legitimately
-            # be missing on a freshly-bootstrapped project — skip them
-            # rather than fail the whole `fetch all` run so the rest of
-            # the pipeline can proceed. Operators inspect the per-source
-            # skip lines and stage the missing inputs as needed.
-            #   - mwbm-climgrid: CAPTCHA-gated ScienceBase download.
-            #   - daymet: pre-staged zarr root not yet configured.
-            #   - margulis-wus-sr: Oregon-only; legitimately absent for
-            #     non-Oregon projects.
-            manual_skip = {"mwbm-climgrid", "daymet", "margulis-wus-sr"}
-            if name in manual_skip:
-                console.print(
-                    f"[yellow]{name}: skipped (manual download not yet "
-                    f"placed; see docs/sources/{name.replace('-', '_')}.md)[/yellow]"
-                )
-                continue
-            print(f"Error fetching {name}: {exc}", file=sys.stderr)
-            sys.exit(1)
-        except (ValueError, RuntimeError) as exc:
-            print(f"Error fetching {name}: {exc}", file=sys.stderr)
-            sys.exit(1)
-        except Exception as exc:
-            _logger.exception("Unexpected error during %s fetch", name)
-            print(
-                f"Unexpected error fetching {name} ({type(exc).__name__}): {exc}",
-                file=sys.stderr,
-            )
-            sys.exit(1)
-
-    console.print(
-        f"\n[bold green]All {len(results)} sources fetched successfully.[/bold green]"
-    )
-
-
-@fetch_app.command(name="merra2")
-def fetch_merra2_cmd(
-    workdir: Annotated[
-        Path,
-        Parameter(
-            name=["--project-dir"],
-            help="Project created by 'nhf-targets init'.",
-        ),
-    ],
-    period: Annotated[
-        str,
-        Parameter(name=["--period", "-p"], help="Temporal range as 'YYYY/YYYY'."),
-    ],
-):
-    """Download MERRA-2 monthly land surface data (M2TMNXLND).
-
-    Authenticates via earthaccess, searches for granules matching the
-    fabric bounding box, downloads them, and prints the provenance record.
-    """
-    import json as json_mod
-
-    from rich.console import Console
-
-    from nhf_spatial_targets.fetch.merra2 import fetch_merra2
-
-    if not workdir.exists():
-        print(f"Error: Project not found: {workdir}", file=sys.stderr)
-        sys.exit(2)
-
-    console = Console()
-    console.print(f"[bold]Fetching MERRA-2 for period {period}...[/bold]")
-
-    try:
-        result = fetch_merra2(workdir=workdir, period=period)
-    except (ValueError, FileNotFoundError, RuntimeError) as exc:
-        print(f"Error: {exc}", file=sys.stderr)
-        sys.exit(1)
-    except Exception as exc:
-        _logger.exception("Unexpected error during MERRA-2 fetch")
-        print(
-            f"Unexpected error ({type(exc).__name__}): {exc}",
-            file=sys.stderr,
-        )
-        sys.exit(1)
-
-    console.print("[green]MERRA-2: downloaded to datastore[/green]")
-    console.print(json_mod.dumps(result, indent=2))
-
-
-@fetch_app.command(name="nldas-mosaic")
-def fetch_nldas_mosaic_cmd(
-    workdir: Annotated[
-        Path,
-        Parameter(
-            name=["--project-dir"],
-            help="Project created by 'nhf-targets init'.",
-        ),
-    ],
-    period: Annotated[
-        str,
-        Parameter(name=["--period", "-p"], help="Temporal range as 'YYYY/YYYY'."),
-    ],
-):
-    """Download NLDAS-2 MOSAIC soil moisture data."""
-    import json as json_mod
-
-    from rich.console import Console
-
-    from nhf_spatial_targets.fetch.nldas import fetch_nldas_mosaic
-
-    if not workdir.exists():
-        print(f"Error: Project not found: {workdir}", file=sys.stderr)
-        sys.exit(2)
-
-    console = Console()
-    console.print(f"[bold]Fetching NLDAS-2 MOSAIC for period {period}...[/bold]")
-
-    try:
-        result = fetch_nldas_mosaic(workdir=workdir, period=period)
-    except (ValueError, FileNotFoundError, RuntimeError) as exc:
-        print(f"Error: {exc}", file=sys.stderr)
-        sys.exit(1)
-    except Exception as exc:
-        _logger.exception("Unexpected error during NLDAS-2 MOSAIC fetch")
-        print(
-            f"Unexpected error ({type(exc).__name__}): {exc}",
-            file=sys.stderr,
-        )
-        sys.exit(1)
-
-    console.print("[green]NLDAS-2 MOSAIC: downloaded to datastore[/green]")
-    console.print(json_mod.dumps(result, indent=2))
-
-
-@fetch_app.command(name="nldas-noah")
-def fetch_nldas_noah_cmd(
-    workdir: Annotated[
-        Path,
-        Parameter(
-            name=["--project-dir"],
-            help="Project created by 'nhf-targets init'.",
-        ),
-    ],
-    period: Annotated[
-        str,
-        Parameter(name=["--period", "-p"], help="Temporal range as 'YYYY/YYYY'."),
-    ],
-):
-    """Download NLDAS-2 NOAH soil moisture data."""
-    import json as json_mod
-
-    from rich.console import Console
-
-    from nhf_spatial_targets.fetch.nldas import fetch_nldas_noah
-
-    if not workdir.exists():
-        print(f"Error: Project not found: {workdir}", file=sys.stderr)
-        sys.exit(2)
-
-    console = Console()
-    console.print(f"[bold]Fetching NLDAS-2 NOAH for period {period}...[/bold]")
-
-    try:
-        result = fetch_nldas_noah(workdir=workdir, period=period)
-    except (ValueError, FileNotFoundError, RuntimeError) as exc:
-        print(f"Error: {exc}", file=sys.stderr)
-        sys.exit(1)
-    except Exception as exc:
-        _logger.exception("Unexpected error during NLDAS-2 NOAH fetch")
-        print(
-            f"Unexpected error ({type(exc).__name__}): {exc}",
-            file=sys.stderr,
-        )
-        sys.exit(1)
-
-    console.print("[green]NLDAS-2 NOAH: downloaded to datastore[/green]")
-    console.print(json_mod.dumps(result, indent=2))
-
-
-@fetch_app.command(name="ncep-ncar")
-def fetch_ncep_ncar_cmd(
-    workdir: Annotated[
-        Path,
-        Parameter(
-            name=["--project-dir"],
-            help="Project created by 'nhf-targets init'.",
-        ),
-    ],
-    period: Annotated[
-        str,
-        Parameter(name=["--period", "-p"], help="Temporal range as 'YYYY/YYYY'."),
-    ],
-):
-    """Download NCEP/NCAR Reanalysis soil moisture data."""
-    import json as json_mod
-
-    from rich.console import Console
-
-    from nhf_spatial_targets.fetch.ncep_ncar import fetch_ncep_ncar
-
-    if not workdir.exists():
-        print(f"Error: Project not found: {workdir}", file=sys.stderr)
-        sys.exit(2)
-
-    console = Console()
-    console.print(f"[bold]Fetching NCEP/NCAR Reanalysis for period {period}...[/bold]")
-
-    try:
-        result = fetch_ncep_ncar(workdir=workdir, period=period)
-    except (ValueError, FileNotFoundError, RuntimeError) as exc:
-        print(f"Error: {exc}", file=sys.stderr)
-        sys.exit(1)
-    except Exception as exc:
-        _logger.exception("Unexpected error during NCEP/NCAR fetch")
-        print(
-            f"Unexpected error ({type(exc).__name__}): {exc}",
-            file=sys.stderr,
-        )
-        sys.exit(1)
-
-    console.print("[green]NCEP/NCAR: downloaded to datastore[/green]")
-    console.print(json_mod.dumps(result, indent=2))
-
-
-@fetch_app.command(name="mod16a2")
-def fetch_mod16a2_cmd(
-    workdir: Annotated[
-        Path,
-        Parameter(
-            name=["--project-dir"],
-            help="Project created by 'nhf-targets init'.",
-        ),
-    ],
-    period: Annotated[
-        str,
-        Parameter(name=["--period", "-p"], help="Temporal range as 'YYYY/YYYY'."),
-    ],
-    force: Annotated[
-        bool,
-        Parameter(
-            name=["--force"],
-            help=(
-                "Re-fetch every year in --period, ignoring the manifest's "
-                "year-skip. Use after a pipeline change has invalidated the "
-                "existing consolidated NCs (e.g. PR #88's fill-mask fix). The "
-                "manifest entry for mod16a2_v061 is overwritten when the run "
-                "completes."
-            ),
-        ),
-    ] = False,
-):
-    """Download MODIS MOD16A2 v061 AET data (8-day composites, 500m)."""
-    import json as json_mod
-
-    from rich.console import Console
-
-    from nhf_spatial_targets.fetch.modis import fetch_mod16a2
-
-    if not workdir.exists():
-        print(f"Error: Project not found: {workdir}", file=sys.stderr)
-        sys.exit(2)
-
-    console = Console()
-    if force:
-        console.print(
-            f"[bold]Force re-fetching MOD16A2 v061 for period {period}...[/bold]"
-        )
-    else:
-        console.print(f"[bold]Fetching MOD16A2 v061 for period {period}...[/bold]")
-
-    try:
-        result = fetch_mod16a2(workdir=workdir, period=period, force=force)
-    except (ValueError, FileNotFoundError, RuntimeError) as exc:
-        print(f"Error: {exc}", file=sys.stderr)
-        sys.exit(1)
-    except Exception as exc:
-        _logger.exception("Unexpected error during MOD16A2 fetch")
-        print(
-            f"Unexpected error ({type(exc).__name__}): {exc}",
-            file=sys.stderr,
-        )
-        sys.exit(1)
-
-    console.print("[green]MOD16A2: downloaded to datastore[/green]")
-    console.print(json_mod.dumps(result, indent=2))
-
-
-@fetch_app.command(name="mod10c1")
-def fetch_mod10c1_cmd(
-    workdir: Annotated[
-        Path,
-        Parameter(
-            name=["--project-dir"],
-            help="Project created by 'nhf-targets init'.",
-        ),
-    ],
-    period: Annotated[
-        str,
-        Parameter(name=["--period", "-p"], help="Temporal range as 'YYYY/YYYY'."),
-    ],
-    force: Annotated[
-        bool,
-        Parameter(
-            name=["--force"],
-            help=(
-                "Re-fetch every year in --period, ignoring the manifest's "
-                "year-skip. The manifest entry for mod10c1_v061 is "
-                "overwritten when the run completes."
-            ),
-        ),
-    ] = False,
-):
-    """Download MODIS MOD10C1 v061 daily snow cover data (0.05deg CMG)."""
-    import json as json_mod
-
-    from rich.console import Console
-
-    from nhf_spatial_targets.fetch.modis import fetch_mod10c1
-
-    if not workdir.exists():
-        print(f"Error: Project not found: {workdir}", file=sys.stderr)
-        sys.exit(2)
-
-    console = Console()
-    if force:
-        console.print(
-            f"[bold]Force re-fetching MOD10C1 v061 for period {period}...[/bold]"
-        )
-    else:
-        console.print(f"[bold]Fetching MOD10C1 v061 for period {period}...[/bold]")
-
-    try:
-        result = fetch_mod10c1(workdir=workdir, period=period, force=force)
-    except (ValueError, FileNotFoundError, RuntimeError) as exc:
-        print(f"Error: {exc}", file=sys.stderr)
-        sys.exit(1)
-    except Exception as exc:
-        _logger.exception("Unexpected error during MOD10C1 fetch")
-        print(
-            f"Unexpected error ({type(exc).__name__}): {exc}",
-            file=sys.stderr,
-        )
-        sys.exit(1)
-
-    console.print("[green]MOD10C1: downloaded to datastore[/green]")
-    console.print(json_mod.dumps(result, indent=2))
-
-
-@fetch_app.command(name="watergap22d")
-def fetch_watergap22d_cmd(
-    workdir: Annotated[
-        Path,
-        Parameter(
-            name=["--project-dir"],
-            help="Project created by 'nhf-targets init'.",
-        ),
-    ],
-    period: Annotated[
-        str,
-        Parameter(name=["--period", "-p"], help="Temporal range as 'YYYY/YYYY'."),
-    ],
-):
-    """Download WaterGAP 2.2d groundwater recharge from PANGAEA."""
-    import json as json_mod
-
-    from rich.console import Console
-
-    from nhf_spatial_targets.fetch.pangaea import fetch_watergap22d
-
-    if not workdir.exists():
-        print(f"Error: Project not found: {workdir}", file=sys.stderr)
-        sys.exit(2)
-
-    console = Console()
-    console.print(f"[bold]Fetching WaterGAP 2.2d for period {period}...[/bold]")
-
-    try:
-        result = fetch_watergap22d(workdir=workdir, period=period)
-    except (ValueError, FileNotFoundError, RuntimeError) as exc:
-        print(f"Error: {exc}", file=sys.stderr)
-        sys.exit(1)
-    except Exception as exc:
-        _logger.exception("Unexpected error during WaterGAP 2.2d fetch")
-        print(
-            f"Unexpected error ({type(exc).__name__}): {exc}",
-            file=sys.stderr,
-        )
-        sys.exit(1)
-
-    console.print("[green]WaterGAP 2.2d: downloaded to datastore[/green]")
-    console.print(json_mod.dumps(result, indent=2))
-
-
-@fetch_app.command(name="era5-land")
-def fetch_era5_land_cmd(
-    workdir: Annotated[
-        Path,
-        Parameter(
-            name=["--project-dir"],
-            help="Project created by 'nhf-targets init'.",
-        ),
-    ],
-    period: Annotated[
-        str,
-        Parameter(name=["--period", "-p"], help="Temporal range as 'YYYY/YYYY'."),
-    ] = "1979/2024",
-    worker_index: Annotated[
-        int,
-        Parameter(
-            name=["--worker-index"],
-            help="0-based index of this worker within the pool (default 0). "
-            "Set to $SLURM_ARRAY_TASK_ID in array jobs.",
-        ),
-    ] = 0,
-    n_workers: Annotated[
-        int,
-        Parameter(
-            name=["--n-workers"],
-            help="Total number of parallel workers (default 1 = serial). "
-            "Must match the SLURM array size.",
-        ),
-    ] = 1,
-):
-    """Download ERA5-Land hourly fields (ro, sro, ssro, sd) via CDS API and consolidate to daily/monthly NetCDFs.
-
-    The runoff variables (ro/sro/ssro) are accumulated and aggregated as
-    daily/monthly sums; sd (snow depth water equivalent) is instantaneous
-    and aggregated as daily/monthly means.
-    """
-    import json as json_mod
-
-    from rich.console import Console
-
-    from nhf_spatial_targets.fetch.era5_land import fetch_era5_land
-
-    if not workdir.exists():
-        print(f"Error: Project not found: {workdir}", file=sys.stderr)
-        sys.exit(2)
-
-    console = Console()
-    if n_workers > 1:
-        console.print(
-            f"[bold]Fetching ERA5-Land for period {period} "
-            f"(worker {worker_index}/{n_workers})...[/bold]"
-        )
-    else:
-        console.print(f"[bold]Fetching ERA5-Land for period {period}...[/bold]")
-
-    try:
-        result = fetch_era5_land(
-            workdir=workdir,
-            period=period,
-            worker_index=worker_index,
-            n_workers=n_workers,
-        )
-    except (ValueError, FileNotFoundError, RuntimeError) as exc:
-        print(f"Error: {exc}", file=sys.stderr)
-        sys.exit(1)
-    except Exception as exc:
-        _logger.exception("Unexpected error during ERA5-Land fetch")
-        print(
-            f"Unexpected error ({type(exc).__name__}): {exc}",
-            file=sys.stderr,
-        )
-        sys.exit(1)
-
-    console.print("[green]ERA5-Land: downloaded to datastore[/green]")
-    console.print(json_mod.dumps(result, indent=2))
-
-
-@fetch_app.command(name="gldas")
-def fetch_gldas_cmd(
-    workdir: Annotated[
-        Path,
-        Parameter(
-            name=["--project-dir"],
-            help="Project created by 'nhf-targets init'.",
-        ),
-    ],
-    period: Annotated[
-        str,
-        Parameter(name=["--period", "-p"], help="Temporal range as 'YYYY/YYYY'."),
-    ] = "2000/2023",
-):
-    """Download GLDAS-2 monthly runoff data via NASA earthaccess."""
-    import json as json_mod
-
-    from rich.console import Console
-
-    from nhf_spatial_targets.fetch.gldas import fetch_gldas
-
-    if not workdir.exists():
-        print(f"Error: Project not found: {workdir}", file=sys.stderr)
-        sys.exit(2)
-
-    console = Console()
-    console.print(f"[bold]Fetching GLDAS for period {period}...[/bold]")
-
-    try:
-        result = fetch_gldas(workdir=workdir, period=period)
-    except (ValueError, FileNotFoundError, RuntimeError) as exc:
-        print(f"Error: {exc}", file=sys.stderr)
-        sys.exit(1)
-    except Exception as exc:
-        _logger.exception("Unexpected error during GLDAS fetch")
-        print(
-            f"Unexpected error ({type(exc).__name__}): {exc}",
-            file=sys.stderr,
-        )
-        sys.exit(1)
-
-    console.print("[green]GLDAS: downloaded to datastore[/green]")
-    console.print(json_mod.dumps(result, indent=2))
-
-
-@fetch_app.command(name="reitz2017")
-def fetch_reitz2017_cmd(
-    workdir: Annotated[
-        Path,
-        Parameter(
-            name=["--project-dir"],
-            help="Project created by 'nhf-targets init'.",
-        ),
-    ],
-    period: Annotated[
-        str,
-        Parameter(name=["--period", "-p"], help="Temporal range as 'YYYY/YYYY'."),
-    ],
-):
-    """Download Reitz 2017 annual recharge estimates from USGS ScienceBase."""
-    import json as json_mod
-
-    from rich.console import Console
-
-    from nhf_spatial_targets.fetch.reitz2017 import fetch_reitz2017
-
-    if not workdir.exists():
-        print(f"Error: Project not found: {workdir}", file=sys.stderr)
-        sys.exit(2)
-
-    console = Console()
-    console.print(f"[bold]Fetching Reitz 2017 recharge for period {period}...[/bold]")
-
-    try:
-        result = fetch_reitz2017(workdir=workdir, period=period)
-    except (ValueError, FileNotFoundError, RuntimeError) as exc:
-        print(f"Error: {exc}", file=sys.stderr)
-        sys.exit(1)
-    except Exception as exc:
-        _logger.exception("Unexpected error during Reitz 2017 fetch")
-        print(
-            f"Unexpected error ({type(exc).__name__}): {exc}",
-            file=sys.stderr,
-        )
-        sys.exit(1)
-
-    console.print("[green]Reitz 2017: downloaded to datastore[/green]")
-    console.print(json_mod.dumps(result, indent=2))
-
-
-@fetch_app.command(name="mwbm-climgrid")
-def fetch_mwbm_climgrid_cmd(
-    workdir: Annotated[
-        Path,
-        Parameter(
-            name=["--project-dir"],
-            help="Project created by 'nhf-targets init'.",
-        ),
-    ],
-    period: Annotated[
-        str,
-        Parameter(name=["--period", "-p"], help="Temporal range as 'YYYY/YYYY'."),
-    ],
-):
-    """Register a manually-placed USGS MWBM (ClimGrid-forced) NetCDF.
-
-    The ScienceBase distribution is gated by a CAPTCHA, so the ~7.5 GB
-    ClimGrid_WBM.nc cannot be retrieved automatically. Download it via
-    a browser and place it at <datastore>/mwbm_climgrid/ClimGrid_WBM.nc
-    before invoking this command — see docs/sources/mwbm_climgrid.md
-    for the procedure. This command then fingerprints the file and
-    writes its provenance to manifest.json.
-    """
-    import json as json_mod
-
-    from rich.console import Console
-
-    from nhf_spatial_targets.fetch.mwbm_climgrid import fetch_mwbm_climgrid
-
-    if not workdir.exists():
-        print(f"Error: Project not found: {workdir}", file=sys.stderr)
-        sys.exit(2)
-
-    console = Console()
-    console.print(f"[bold]Registering MWBM ClimGrid (period {period})...[/bold]")
-
-    try:
-        result = fetch_mwbm_climgrid(workdir=workdir, period=period)
-    except (ValueError, FileNotFoundError, RuntimeError) as exc:
-        print(f"Error: {exc}", file=sys.stderr)
-        sys.exit(1)
-    except Exception as exc:
-        _logger.exception("Unexpected error during MWBM ClimGrid registration")
-        print(
-            f"Unexpected error ({type(exc).__name__}): {exc}",
-            file=sys.stderr,
-        )
-        sys.exit(1)
-
-    console.print("[green]MWBM ClimGrid: registered in manifest[/green]")
-    console.print(json_mod.dumps(result, indent=2))
-
-
-# ---------------------------------------------------------------------------
-# SWE category (issue #99) — fetch scaffolding
-# Aggregate + target wiring deferred to per-source follow-up issues.
-# ---------------------------------------------------------------------------
-
-
-@fetch_app.command(name="daymet")
-def fetch_daymet_cmd(
-    workdir: Annotated[
-        Path,
-        Parameter(
-            name=["--project-dir"],
-            help="Project created by 'nhf-targets init'.",
-        ),
-    ],
-    period: Annotated[
-        str,
-        Parameter(name=["--period", "-p"], help="Temporal range as 'YYYY/YYYY'."),
-    ] = "1980/2024",
-    source_path: Annotated[
-        Path | None,
-        Parameter(
-            name=["--source-path"],
-            help="Directory containing daymet_{na,hi,pr}.zarr. Overrides "
-            "'daymet_root' from config.yml.",
-        ),
-    ] = None,
-    region: Annotated[
-        str,
-        Parameter(
-            name=["--region"],
-            help="One of 'na' | 'hi' | 'pr' | 'all' (default: all).",
-        ),
-    ] = "all",
-):
-    """Register operator-staged Daymet V4 R1 regional zarr stores.
-
-    Daymet zarrs are pre-staged on a shared filesystem; this command
-    fingerprints the structural metadata of each region and records a
-    per-region manifest entry. No downloading or copying. See
-    docs/sources/daymet.md for the staging procedure.
-    """
-    import json as json_mod
-
-    from rich.console import Console
-
-    from nhf_spatial_targets.fetch.daymet import fetch_daymet
-
-    if not workdir.exists():
-        print(f"Error: Project not found: {workdir}", file=sys.stderr)
-        sys.exit(2)
-
-    console = Console()
-    console.print(
-        f"[bold]Registering Daymet (region={region}, period {period})...[/bold]"
-    )
-
-    try:
-        result = fetch_daymet(
-            workdir=workdir,
-            period=period,
-            source_path=source_path,
-            region=region,
-        )
-    except (ValueError, FileNotFoundError, RuntimeError) as exc:
-        print(f"Error: {exc}", file=sys.stderr)
-        sys.exit(1)
-    except Exception as exc:
-        _logger.exception("Unexpected error during Daymet registration")
-        print(
-            f"Unexpected error ({type(exc).__name__}): {exc}",
-            file=sys.stderr,
-        )
-        sys.exit(1)
-
-    console.print("[green]Daymet: registered in manifest[/green]")
-    console.print(json_mod.dumps(result, indent=2))
-
-
-@fetch_app.command(name="snodas")
-def fetch_snodas_cmd(
-    workdir: Annotated[
-        Path,
-        Parameter(
-            name=["--project-dir"],
-            help="Project created by 'nhf-targets init'.",
-        ),
-    ],
-    period: Annotated[
-        str,
-        Parameter(name=["--period", "-p"], help="Temporal range as 'YYYY/YYYY'."),
-    ] = "2003/2024",
-    worker_index: Annotated[
-        int,
-        Parameter(
-            name=["--worker-index"],
-            help="0-based index of this worker within the pool (default 0).",
-        ),
-    ] = 0,
-    n_workers: Annotated[
-        int,
-        Parameter(
-            name=["--n-workers"],
-            help="Total number of parallel workers (default 1 = serial).",
-        ),
-    ] = 1,
-):
-    """Download SNODAS daily SWE granules from NSIDC (G02158) via earthaccess.
-
-    Fetch-only: raw tar/.Hdr bundles land in <datastore>/snodas/raw/<year>/.
-    Decoding into CF NetCDFs is deferred to the SNODAS aggregate follow-up.
-    """
-    import json as json_mod
-
-    from rich.console import Console
-
-    from nhf_spatial_targets.fetch.snodas import fetch_snodas
-
-    if not workdir.exists():
-        print(f"Error: Project not found: {workdir}", file=sys.stderr)
-        sys.exit(2)
-
-    console = Console()
-    if n_workers > 1:
-        console.print(
-            f"[bold]Fetching SNODAS for period {period} "
-            f"(worker {worker_index}/{n_workers})...[/bold]"
-        )
-    else:
-        console.print(f"[bold]Fetching SNODAS for period {period}...[/bold]")
-
-    try:
-        result = fetch_snodas(
-            workdir=workdir,
-            period=period,
-            worker_index=worker_index,
-            n_workers=n_workers,
-        )
-    except (ValueError, FileNotFoundError, RuntimeError) as exc:
-        print(f"Error: {exc}", file=sys.stderr)
-        sys.exit(1)
-    except Exception as exc:
-        _logger.exception("Unexpected error during SNODAS fetch")
-        print(
-            f"Unexpected error ({type(exc).__name__}): {exc}",
-            file=sys.stderr,
-        )
-        sys.exit(1)
-
-    console.print("[green]SNODAS: downloaded to datastore[/green]")
-    console.print(json_mod.dumps(result, indent=2))
-
-
-@fetch_app.command(name="margulis-wus-sr")
-def fetch_margulis_wus_sr_cmd(
-    workdir: Annotated[
-        Path,
-        Parameter(
-            name=["--project-dir"],
-            help="Project created by 'nhf-targets init'.",
-        ),
-    ],
-    period: Annotated[
-        str,
-        Parameter(name=["--period", "-p"], help="Temporal range as 'YYYY/YYYY'."),
-    ] = "1985/2021",
-):
-    """Download Margulis Western US Snow Reanalysis (NSIDC-0719) via earthaccess.
-
-    Fabric-scoped to Oregon only (catalog `fabric_scope`); the scope is
-    recorded in manifest.json but not enforced at fetch time. Fetch-only:
-    consolidation is deferred to the Margulis aggregate follow-up.
-    """
-    import json as json_mod
-
-    from rich.console import Console
-
-    from nhf_spatial_targets.fetch.margulis_wus_sr import fetch_margulis_wus_sr
-
-    if not workdir.exists():
-        print(f"Error: Project not found: {workdir}", file=sys.stderr)
-        sys.exit(2)
-
-    console = Console()
-    console.print(f"[bold]Fetching Margulis WUS-SR for period {period}...[/bold]")
-
-    try:
-        result = fetch_margulis_wus_sr(workdir=workdir, period=period)
-    except (ValueError, FileNotFoundError, RuntimeError) as exc:
-        print(f"Error: {exc}", file=sys.stderr)
-        sys.exit(1)
-    except Exception as exc:
-        _logger.exception("Unexpected error during Margulis WUS-SR fetch")
-        print(
-            f"Unexpected error ({type(exc).__name__}): {exc}",
-            file=sys.stderr,
-        )
-        sys.exit(1)
-
-    console.print("[green]Margulis WUS-SR: downloaded to datastore[/green]")
-    console.print(json_mod.dumps(result, indent=2))
-
-
-@agg_app.command(name="ssebop")
-def agg_ssebop_cmd(
-    workdir: Annotated[
-        Path,
-        Parameter(
-            name=["--project-dir"],
-            help="Project created by 'nhf-targets init'.",
-        ),
-    ],
-    period: Annotated[
-        str,
-        Parameter(name=["--period", "-p"], help="Temporal range as 'YYYY/YYYY'."),
-    ],
-    batch_size: Annotated[int | None, _AGG_BATCH_SIZE_PARAM] = None,
-    worker_index: Annotated[int, _AGG_WORKER_INDEX_PARAM] = 0,
-    n_workers: Annotated[int, _AGG_N_WORKERS_PARAM] = 1,
-):
-    """Aggregate SSEBop monthly AET to HRU fabric polygons.
-
-    Reads SSEBop data from the USGS NHGF STAC catalog (Zarr), computes
-    area-weighted means per HRU, and writes the result to NetCDF.
-    """
-    from rich.console import Console
-
-    from nhf_spatial_targets.aggregate.ssebop import aggregate_ssebop
-
-    if not workdir.exists():
-        print(f"Error: Project not found: {workdir}", file=sys.stderr)
-        sys.exit(2)
-    if not (workdir / "fabric.json").exists():
-        print(
-            f"Error: fabric.json not found in {workdir}. "
-            "Run 'nhf-targets validate' first.",
-            file=sys.stderr,
-        )
-        sys.exit(2)
-
-    fabric_path, id_col, resolved_batch_size = _resolve_agg_config(workdir, batch_size)
-
-    console = Console()
-    worker_suffix = f", worker={worker_index}/{n_workers}" if n_workers > 1 else ""
-    console.print(
-        f"[bold]Aggregating SSEBop AET for period {period} "
-        f"(batch_size={resolved_batch_size}{worker_suffix})...[/bold]"
-    )
-
-    try:
-        aggregate_ssebop(
-            fabric_path=fabric_path,
-            id_col=id_col,
-            period=period,
-            workdir=workdir,
-            batch_size=resolved_batch_size,
-            worker_index=worker_index,
-            n_workers=n_workers,
-        )
-    except (ValueError, FileNotFoundError, RuntimeError) as exc:
-        print(f"Error: {exc}", file=sys.stderr)
-        sys.exit(1)
-    except Exception as exc:
-        _logger.exception("Unexpected error during SSEBop aggregation")
-        print(
-            f"Unexpected error ({type(exc).__name__}): {exc}",
-            file=sys.stderr,
-        )
-        sys.exit(1)
-
-    console.print(
-        f"[green]SSEBop aggregation complete; per-year NCs and manifest "
-        f"updated under {workdir / 'data' / 'aggregated' / 'ssebop'}[/green]"
-    )
-
-
-@agg_app.command(name="daymet")
-def agg_daymet_cmd(
-    workdir: Annotated[
-        Path,
-        Parameter(
-            name=["--project-dir"],
-            help="Project created by 'nhf-targets init'.",
-        ),
-    ],
-    period: Annotated[
-        str,
-        Parameter(name=["--period", "-p"], help="Temporal range as 'YYYY/YYYY'."),
-    ],
-    batch_size: Annotated[int | None, _AGG_BATCH_SIZE_PARAM] = None,
-    worker_index: Annotated[int, _AGG_WORKER_INDEX_PARAM] = 0,
-    n_workers: Annotated[int, _AGG_N_WORKERS_PARAM] = 1,
-    region: Annotated[
-        str,
-        Parameter(
-            name=["--region", "-r"],
-            help=(
-                "Daymet region: 'na', 'hi', or 'pr'. Only 'na' is wired "
-                "up in this build; 'hi' and 'pr' raise NotImplementedError "
-                "until the corresponding fabrics land (issue #101)."
-            ),
-        ),
-    ] = "na",
-):
-    """Aggregate Daymet V4 R1 daily SWE to HRU fabric polygons.
-
-    Reads the per-region zarr path from manifest.json (written by
-    'nhf-targets fetch daymet'), opens the zarr directly, computes
-    area-weighted means per HRU, and writes per-year NetCDFs.
-    """
-    from rich.console import Console
-
-    if not workdir.exists():
-        print(f"Error: Project not found: {workdir}", file=sys.stderr)
-        sys.exit(2)
-    if not (workdir / "fabric.json").exists():
-        print(
-            f"Error: fabric.json not found in {workdir}. "
-            "Run 'nhf-targets validate' first.",
-            file=sys.stderr,
-        )
-        sys.exit(2)
-
-    fabric_path, id_col, resolved_batch_size = _resolve_agg_config(workdir, batch_size)
-
-    console = Console()
-    worker_suffix = f", worker={worker_index}/{n_workers}" if n_workers > 1 else ""
-    console.print(
-        f"[bold]Aggregating Daymet SWE (region={region}, period={period}, "
-        f"batch_size={resolved_batch_size}{worker_suffix})...[/bold]"
-    )
-
-    try:
-        aggregate_daymet(
-            fabric_path=fabric_path,
-            id_col=id_col,
-            period=period,
-            workdir=workdir,
-            batch_size=resolved_batch_size,
-            region=region,
-            worker_index=worker_index,
-            n_workers=n_workers,
-        )
-    except (ValueError, FileNotFoundError, RuntimeError, NotImplementedError) as exc:
-        print(f"Error ({type(exc).__name__}): {exc}", file=sys.stderr)
-        sys.exit(1)
-    except Exception as exc:
-        _logger.exception("Unexpected error during Daymet aggregation")
-        print(
-            f"Unexpected error ({type(exc).__name__}): {exc}",
-            file=sys.stderr,
-        )
-        sys.exit(1)
-
-    console.print(
-        f"[green]Daymet aggregation complete; per-year NCs and manifest "
-        f"updated under {workdir / 'data' / 'aggregated' / 'daymet'}[/green]"
-    )
-
-
-@catalog_app.command(name="sources")
-def catalog_sources():
-    """List all registered data sources."""
-    from nhf_spatial_targets.catalog import sources
-    from rich import print as rprint
-
-    rprint(sources())
-
-
-@catalog_app.command(name="variables")
-def catalog_variables():
-    """List all calibration variable definitions."""
-    from nhf_spatial_targets.catalog import variables
-    from rich import print as rprint
-
-    rprint(variables())
-
-
-def _resolve_agg_config(
-    workdir: Path, cli_batch_size: int | None
-) -> tuple[str, str, int]:
-    """Return (fabric_path, id_col, batch_size) from project config.
-
-    ``batch_size`` resolution order: CLI ``--batch-size`` flag (if non-None)
-    > ``config.yml[fabric.batch_size]`` > 500 (defaults.py).
-    """
-    from nhf_spatial_targets.defaults import apply_defaults
-
-    try:
-        user_cfg = yaml.safe_load((workdir / "config.yml").read_text())
-    except yaml.YAMLError as exc:
-        print(f"Error: Cannot parse config.yml: {exc}", file=sys.stderr)
-        sys.exit(1)
-    cfg = apply_defaults(user_cfg)
-    fabric_cfg = cfg.get("fabric") or {}
-    fabric_path = fabric_cfg["path"]
-    id_col = fabric_cfg.get("id_col", "nhm_id")
-    if cli_batch_size is not None:
-        batch_size = int(cli_batch_size)
-    else:
-        batch_size = int(fabric_cfg.get("batch_size", 500))
-    return fabric_path, id_col, batch_size
-
-
-def _run_tier_agg(
-    aggregate_fn,
-    label: str,
-    workdir: Path,
-    batch_size: int | None,
-    period: str | None = None,
-    *,
-    worker_index: int = 0,
-    n_workers: int = 1,
-) -> None:
-    """Common boilerplate for tier-1/tier-2 aggregator CLI wrappers.
-
-    ``period`` is forwarded to ``aggregate_fn`` only when set, so
-    aggregators that don't accept it (most sources, where fetch already
-    clips by file) are unaffected.
-
-    ``batch_size`` of ``None`` means "fall back to ``config.yml``" via
-    :func:`_resolve_agg_config`. The CLI per-command default is now
-    ``None`` so projects can pin batch_size in their config rather than
-    relying on every operator passing ``--batch-size 500`` (issue #156).
-
-    ``worker_index`` / ``n_workers`` enable SLURM-array year-sharding
-    (issue #156). Default ``(0, 1)`` is the single-worker serial path.
-    """
-    from rich.console import Console
-
-    if not workdir.exists():
-        print(f"Error: Project not found: {workdir}", file=sys.stderr)
-        sys.exit(2)
-    if not (workdir / "fabric.json").exists():
-        print(
-            f"Error: fabric.json not found in {workdir}. "
-            "Run 'nhf-targets validate' first.",
-            file=sys.stderr,
-        )
-        sys.exit(2)
-
-    fabric_path, id_col, resolved_batch_size = _resolve_agg_config(workdir, batch_size)
-
-    console = Console()
-    period_suffix = f", period={period}" if period is not None else ""
-    worker_suffix = f", worker={worker_index}/{n_workers}" if n_workers > 1 else ""
-    console.print(
-        f"[bold]Aggregating {label} (batch_size={resolved_batch_size}"
-        f"{period_suffix}{worker_suffix})...[/bold]"
-    )
-    try:
-        kwargs = {
-            "fabric_path": fabric_path,
-            "id_col": id_col,
-            "workdir": workdir,
-            "batch_size": resolved_batch_size,
-            "worker_index": worker_index,
-            "n_workers": n_workers,
-        }
-        if period is not None:
-            kwargs["period"] = period
-        aggregate_fn(**kwargs)
-    except (ValueError, FileNotFoundError, RuntimeError) as exc:
-        print(f"Error ({type(exc).__name__}): {exc}", file=sys.stderr)
-        sys.exit(1)
-    except Exception as exc:
-        _logger.exception("Unexpected error during %s aggregation", label)
-        print(
-            f"Unexpected error ({type(exc).__name__}): {exc}",
-            file=sys.stderr,
-        )
-        sys.exit(1)
-    console.print(
-        f"[green]{label} aggregation complete; per-year NCs and manifest "
-        f"updated under {workdir}/data/aggregated/[/green]"
-    )
-
-
-@agg_app.command(name="era5-land")
-def agg_era5_land_cmd(
-    workdir: Annotated[Path, Parameter(name=["--project-dir"])],
-    batch_size: Annotated[int | None, _AGG_BATCH_SIZE_PARAM] = None,
-    worker_index: Annotated[int, _AGG_WORKER_INDEX_PARAM] = 0,
-    n_workers: Annotated[int, _AGG_N_WORKERS_PARAM] = 1,
-):
-    """Aggregate ERA5-Land monthly runoff to HRU polygons."""
-    _run_tier_agg(
-        aggregate_era5_land,
-        "ERA5-Land",
-        workdir,
-        batch_size,
-        worker_index=worker_index,
-        n_workers=n_workers,
-    )
-
-
-@agg_app.command(name="gldas")
-def agg_gldas_cmd(
-    workdir: Annotated[Path, Parameter(name=["--project-dir"])],
-    batch_size: Annotated[int | None, _AGG_BATCH_SIZE_PARAM] = None,
-    worker_index: Annotated[int, _AGG_WORKER_INDEX_PARAM] = 0,
-    n_workers: Annotated[int, _AGG_N_WORKERS_PARAM] = 1,
-):
-    """Aggregate GLDAS-2.1 NOAH monthly runoff to HRU polygons."""
-    _run_tier_agg(
-        aggregate_gldas,
-        "GLDAS",
-        workdir,
-        batch_size,
-        worker_index=worker_index,
-        n_workers=n_workers,
-    )
-
-
-@agg_app.command(name="merra2")
-def agg_merra2_cmd(
-    workdir: Annotated[Path, Parameter(name=["--project-dir"])],
-    batch_size: Annotated[int | None, _AGG_BATCH_SIZE_PARAM] = None,
-    worker_index: Annotated[int, _AGG_WORKER_INDEX_PARAM] = 0,
-    n_workers: Annotated[int, _AGG_N_WORKERS_PARAM] = 1,
-):
-    """Aggregate MERRA-2 monthly soil wetness to HRU polygons."""
-    _run_tier_agg(
-        aggregate_merra2,
-        "MERRA-2",
-        workdir,
-        batch_size,
-        worker_index=worker_index,
-        n_workers=n_workers,
-    )
-
-
-@agg_app.command(name="ncep-ncar")
-def agg_ncep_ncar_cmd(
-    workdir: Annotated[Path, Parameter(name=["--project-dir"])],
-    batch_size: Annotated[int | None, _AGG_BATCH_SIZE_PARAM] = None,
-    worker_index: Annotated[int, _AGG_WORKER_INDEX_PARAM] = 0,
-    n_workers: Annotated[int, _AGG_N_WORKERS_PARAM] = 1,
-):
-    """Aggregate NCEP/NCAR monthly soil moisture to HRU polygons."""
-    _run_tier_agg(
-        aggregate_ncep_ncar,
-        "NCEP/NCAR",
-        workdir,
-        batch_size,
-        worker_index=worker_index,
-        n_workers=n_workers,
-    )
-
-
-@agg_app.command(name="nldas-mosaic")
-def agg_nldas_mosaic_cmd(
-    workdir: Annotated[Path, Parameter(name=["--project-dir"])],
-    batch_size: Annotated[int | None, _AGG_BATCH_SIZE_PARAM] = None,
-    worker_index: Annotated[int, _AGG_WORKER_INDEX_PARAM] = 0,
-    n_workers: Annotated[int, _AGG_N_WORKERS_PARAM] = 1,
-):
-    """Aggregate NLDAS-2 MOSAIC monthly soil moisture to HRU polygons."""
-    _run_tier_agg(
-        aggregate_nldas_mosaic,
-        "NLDAS-MOSAIC",
-        workdir,
-        batch_size,
-        worker_index=worker_index,
-        n_workers=n_workers,
-    )
-
-
-@agg_app.command(name="nldas-noah")
-def agg_nldas_noah_cmd(
-    workdir: Annotated[Path, Parameter(name=["--project-dir"])],
-    batch_size: Annotated[int | None, _AGG_BATCH_SIZE_PARAM] = None,
-    worker_index: Annotated[int, _AGG_WORKER_INDEX_PARAM] = 0,
-    n_workers: Annotated[int, _AGG_N_WORKERS_PARAM] = 1,
-):
-    """Aggregate NLDAS-2 NOAH monthly soil moisture to HRU polygons."""
-    _run_tier_agg(
-        aggregate_nldas_noah,
-        "NLDAS-NOAH",
-        workdir,
-        batch_size,
-        worker_index=worker_index,
-        n_workers=n_workers,
-    )
-
-
-@agg_app.command(name="watergap22d")
-def agg_watergap22d_cmd(
-    workdir: Annotated[Path, Parameter(name=["--project-dir"])],
-    batch_size: Annotated[int | None, _AGG_BATCH_SIZE_PARAM] = None,
-    worker_index: Annotated[int, _AGG_WORKER_INDEX_PARAM] = 0,
-    n_workers: Annotated[int, _AGG_N_WORKERS_PARAM] = 1,
-):
-    """Aggregate WaterGAP 2.2d monthly diffuse recharge to HRU polygons."""
-    _run_tier_agg(
-        aggregate_watergap22d,
-        "WaterGAP 2.2d",
-        workdir,
-        batch_size,
-        worker_index=worker_index,
-        n_workers=n_workers,
-    )
-
-
-@agg_app.command(name="reitz2017")
-def agg_reitz2017_cmd(
-    workdir: Annotated[Path, Parameter(name=["--project-dir"])],
-    batch_size: Annotated[int | None, _AGG_BATCH_SIZE_PARAM] = None,
-    worker_index: Annotated[int, _AGG_WORKER_INDEX_PARAM] = 0,
-    n_workers: Annotated[int, _AGG_N_WORKERS_PARAM] = 1,
-):
-    """Aggregate Reitz 2017 annual recharge to HRU polygons."""
-    _run_tier_agg(
-        aggregate_reitz2017,
-        "Reitz 2017",
-        workdir,
-        batch_size,
-        worker_index=worker_index,
-        n_workers=n_workers,
-    )
-
-
-@agg_app.command(name="mod16a2")
-def agg_mod16a2_cmd(
-    workdir: Annotated[Path, Parameter(name=["--project-dir"])],
-    batch_size: Annotated[int | None, _AGG_BATCH_SIZE_PARAM] = None,
-    worker_index: Annotated[int, _AGG_WORKER_INDEX_PARAM] = 0,
-    n_workers: Annotated[int, _AGG_N_WORKERS_PARAM] = 1,
-):
-    """Aggregate MOD16A2 v061 8-day AET to HRU polygons."""
-    _run_tier_agg(
-        aggregate_mod16a2,
-        "MOD16A2",
-        workdir,
-        batch_size,
-        worker_index=worker_index,
-        n_workers=n_workers,
-    )
-
-
-@agg_app.command(name="mod10c1")
-def agg_mod10c1_cmd(
-    workdir: Annotated[Path, Parameter(name=["--project-dir"])],
-    batch_size: Annotated[int | None, _AGG_BATCH_SIZE_PARAM] = None,
-    worker_index: Annotated[int, _AGG_WORKER_INDEX_PARAM] = 0,
-    n_workers: Annotated[int, _AGG_N_WORKERS_PARAM] = 1,
-):
-    """Aggregate MOD10C1 v061 daily SCA to HRU polygons."""
-    _run_tier_agg(
-        aggregate_mod10c1,
-        "MOD10C1",
-        workdir,
-        batch_size,
-        worker_index=worker_index,
-        n_workers=n_workers,
-    )
-
-
-@agg_app.command(name="snodas")
-def agg_snodas_cmd(
-    workdir: Annotated[Path, Parameter(name=["--project-dir"])],
-    batch_size: Annotated[int | None, _AGG_BATCH_SIZE_PARAM] = None,
-    worker_index: Annotated[int, _AGG_WORKER_INDEX_PARAM] = 0,
-    n_workers: Annotated[int, _AGG_N_WORKERS_PARAM] = 1,
-    period: Annotated[
-        str | None,
-        Parameter(
-            name=["--period", "-p"],
-            help=(
-                "Optional 'YYYY/YYYY' clip applied at agg time. SNODAS "
-                "consolidated NCs span 2003-present; pass e.g. '2003/2020' "
-                "to restrict aggregation. Omit to aggregate every year "
-                "present in the datastore."
-            ),
-        ),
-    ] = None,
-):
-    """Aggregate SNODAS daily SWE to HRU polygons."""
-    _run_tier_agg(
-        aggregate_snodas,
-        "SNODAS",
-        workdir,
-        batch_size,
-        period=period,
-        worker_index=worker_index,
-        n_workers=n_workers,
-    )
-
-
-@agg_app.command(name="era5-land-sd")
-def agg_era5_land_sd_cmd(
-    workdir: Annotated[Path, Parameter(name=["--project-dir"])],
-    batch_size: Annotated[int | None, _AGG_BATCH_SIZE_PARAM] = None,
-    worker_index: Annotated[int, _AGG_WORKER_INDEX_PARAM] = 0,
-    n_workers: Annotated[int, _AGG_N_WORKERS_PARAM] = 1,
-):
-    """Aggregate ERA5-Land daily snow depth water equivalent to HRU polygons.
-
-    Reads from ``<datastore>/era5_land/daily/era5_land_daily_*.nc`` and
-    writes to ``<project>/data/aggregated/era5_land_sd/`` so the daily
-    SWE outputs stay separate from the monthly runoff aggregations under
-    ``era5_land/`` (which the runoff and recharge targets consume).
-    """
-    _run_tier_agg(
-        aggregate_era5_land_sd,
-        "ERA5-Land sd",
-        workdir,
-        batch_size,
-        worker_index=worker_index,
-        n_workers=n_workers,
-    )
-
-
-@agg_app.command(name="margulis-wus-sr")
-def agg_margulis_wus_sr_cmd(
-    workdir: Annotated[Path, Parameter(name=["--project-dir"])],
-    batch_size: Annotated[int | None, _AGG_BATCH_SIZE_PARAM] = None,
-    worker_index: Annotated[int, _AGG_WORKER_INDEX_PARAM] = 0,
-    n_workers: Annotated[int, _AGG_N_WORKERS_PARAM] = 1,
-    period: Annotated[
-        str | None,
-        Parameter(
-            name=["--period", "-p"],
-            help=(
-                "Optional 'YYYY/YYYY' clip applied at agg time. Margulis "
-                "consolidated NCs span 1985-2020 in the datastore; pass "
-                "e.g. '2003/2020' to restrict aggregation to the SWE "
-                "target window. Omit to aggregate every year present."
-            ),
-        ),
-    ] = None,
-):
-    """Aggregate Margulis WUS-SR daily SWE to HRU polygons (Oregon-scoped)."""
-    _run_tier_agg(
-        aggregate_margulis_wus_sr,
-        "Margulis WUS-SR",
-        workdir,
-        batch_size,
-        period=period,
-        worker_index=worker_index,
-        n_workers=n_workers,
-    )
-
-
-@agg_app.command(name="mwbm-climgrid")
-def agg_mwbm_climgrid_cmd(
-    workdir: Annotated[Path, Parameter(name=["--project-dir"])],
-    batch_size: Annotated[int | None, _AGG_BATCH_SIZE_PARAM] = None,
-    worker_index: Annotated[int, _AGG_WORKER_INDEX_PARAM] = 0,
-    n_workers: Annotated[int, _AGG_N_WORKERS_PARAM] = 1,
-    period: Annotated[
-        str | None,
-        Parameter(
-            name=["--period", "-p"],
-            help=(
-                "Optional 'YYYY/YYYY' clip applied to the per-year output. "
-                "ClimGrid_WBM.nc spans 1895-2020 in a single file; pass e.g. "
-                "'1979/2020' to skip the publisher's spinup years and "
-                "anything outside the NHM run window. Omit to aggregate "
-                "every year in the file."
-            ),
-        ),
-    ] = None,
-):
-    """Aggregate USGS MWBM (ClimGrid-forced) monthly outputs to HRU polygons."""
-    _run_tier_agg(
-        aggregate_mwbm_climgrid,
-        "MWBM (ClimGrid)",
-        workdir,
-        batch_size,
-        period=period,
-        worker_index=worker_index,
-        n_workers=n_workers,
-    )
-
-
-@agg_app.command(name="all")
-def agg_all_cmd(
-    workdir: Annotated[
-        Path,
-        Parameter(
-            name=["--project-dir"], help="Project created by 'nhf-targets init'."
-        ),
-    ],
-    batch_size: Annotated[int | None, _AGG_BATCH_SIZE_PARAM] = None,
-    worker_index: Annotated[int, _AGG_WORKER_INDEX_PARAM] = 0,
-    n_workers: Annotated[int, _AGG_N_WORKERS_PARAM] = 1,
-):
-    """Aggregate every registered source for this project.
-
-    Runs tier-1/tier-2 aggregators in sequence; stops on first failure.
-    Sources that require an explicit ``--period`` argument are not
-    included here — run them separately:
-
-    - ``agg ssebop --period YYYY/YYYY``
-    - ``agg daymet --period YYYY/YYYY [--region na]``
-
-    SLURM-array support: when called with ``--n-workers > 1`` every
-    source's aggregator gets the same ``worker_index`` / ``n_workers``
-    pair. Year sharding is per-source, so all 14 aggregators run on
-    this one worker for its slice of each source's years.
-    """
-    from rich.console import Console
-
-    console = Console()
-    if not workdir.exists():
-        print(f"Error: Project not found: {workdir}", file=sys.stderr)
-        sys.exit(2)
-
-    sources: list[tuple[str, Callable[..., None]]] = [
-        ("era5-land", aggregate_era5_land),
-        ("gldas", aggregate_gldas),
-        ("merra2", aggregate_merra2),
-        ("ncep-ncar", aggregate_ncep_ncar),
-        ("nldas-mosaic", aggregate_nldas_mosaic),
-        ("nldas-noah", aggregate_nldas_noah),
-        ("watergap22d", aggregate_watergap22d),
-        ("reitz2017", aggregate_reitz2017),
-        ("mod16a2", aggregate_mod16a2),
-        ("mod10c1", aggregate_mod10c1),
-        ("snodas", aggregate_snodas),
-        ("mwbm-climgrid", aggregate_mwbm_climgrid),
-        ("era5-land-sd", aggregate_era5_land_sd),
-        ("margulis-wus-sr", aggregate_margulis_wus_sr),
-    ]
-    for label, fn in sources:
-        console.print(f"\n[bold]{'─' * 60}[/bold]")
-        _run_tier_agg(
-            fn,
-            label,
-            workdir,
-            batch_size,
-            worker_index=worker_index,
-            n_workers=n_workers,
-        )
-
-    console.print(
-        f"\n[bold green]All {len(sources)} sources aggregated successfully.[/bold green]"
-    )
-
-
-main = app.meta
