+++ src/nhf_spatial_targets/cli/fetch.py
@@ -1,0 +1,916 @@
+"""``nhf-targets fetch`` sub-app: download source datasets to the datastore."""
+
+from __future__ import annotations
+
+import logging
+import sys
+from pathlib import Path
+from typing import Annotated
+
+from cyclopts import App, Parameter
+
+_logger = logging.getLogger(__name__)
+
+fetch_app = App(name="fetch", help="Download source datasets into a project datastore.")
+
+
+@fetch_app.command(name="all")
+def fetch_all_cmd(
+    workdir: Annotated[
+        Path,
+        Parameter(
+            name=["--project-dir"],
+            help="Project created by 'nhf-targets init'.",
+        ),
+    ],
+    period: Annotated[
+        str,
+        Parameter(name=["--period", "-p"], help="Temporal range as 'YYYY/YYYY'."),
+    ],
+):
+    """Download all source datasets into the project datastore.
+
+    Iterates through every registered fetch module in sequence.
+    Stops on the first failure.
+    """
+    from rich.console import Console
+
+    if not workdir.exists():
+        print(f"Error: Project not found: {workdir}", file=sys.stderr)
+        sys.exit(2)
+
+    console = Console()
+
+    # Import all fetch functions
+    import nhf_spatial_targets.catalog as _catalog
+    from nhf_spatial_targets.fetch._period import clamp_period
+    from nhf_spatial_targets.fetch.daymet import fetch_daymet
+    from nhf_spatial_targets.fetch.era5_land import fetch_era5_land
+    from nhf_spatial_targets.fetch.gldas import fetch_gldas
+    from nhf_spatial_targets.fetch.margulis_wus_sr import fetch_margulis_wus_sr
+    from nhf_spatial_targets.fetch.merra2 import fetch_merra2
+    from nhf_spatial_targets.fetch.modis import fetch_mod10c1, fetch_mod16a2
+    from nhf_spatial_targets.fetch.mwbm_climgrid import fetch_mwbm_climgrid
+    from nhf_spatial_targets.fetch.ncep_ncar import fetch_ncep_ncar
+    from nhf_spatial_targets.fetch.nldas import fetch_nldas_mosaic, fetch_nldas_noah
+    from nhf_spatial_targets.fetch.pangaea import fetch_watergap22d
+    from nhf_spatial_targets.fetch.reitz2017 import fetch_reitz2017
+    from nhf_spatial_targets.fetch.snodas import fetch_snodas
+
+    # (display name, catalog source key, fetch function)
+    sources = [
+        ("era5-land", "era5_land", fetch_era5_land),
+        ("gldas", "gldas_noah_v21_monthly", fetch_gldas),
+        ("merra2", "merra2", fetch_merra2),
+        ("nldas-mosaic", "nldas_mosaic", fetch_nldas_mosaic),
+        ("nldas-noah", "nldas_noah", fetch_nldas_noah),
+        ("ncep-ncar", "ncep_ncar", fetch_ncep_ncar),
+        ("mod16a2", "mod16a2_v061", fetch_mod16a2),
+        ("mod10c1", "mod10c1_v061", fetch_mod10c1),
+        ("watergap22d", "watergap22d", fetch_watergap22d),
+        ("reitz2017", "reitz2017", fetch_reitz2017),
+        ("mwbm-climgrid", "mwbm_climgrid", fetch_mwbm_climgrid),
+        # SWE category (issue #99) — fetch scaffolding only; aggregate +
+        # target wiring deferred to per-source follow-up issues.
+        ("daymet", "daymet", fetch_daymet),
+        ("snodas", "snodas", fetch_snodas),
+        ("margulis-wus-sr", "margulis_wus_sr", fetch_margulis_wus_sr),
+    ]
+
+    results = {}
+    for name, source_key, fetch_fn in sources:
+        console.print(f"\n[bold]{'─' * 60}[/bold]")
+
+        # Clamp requested period to the source's available range
+        meta = _catalog.source(source_key)
+        available = meta.get("period", "")
+        clamped = clamp_period(period, available) if available else period
+
+        if clamped is None:
+            console.print(
+                f"[yellow]{name}: skipped (no overlap between "
+                f"requested {period} and available {available})[/yellow]"
+            )
+            continue
+
+        if clamped != period:
+            console.print(
+                f"[bold]Fetching {name} for period {clamped} "
+                f"(clamped from {period} to available {available})...[/bold]"
+            )
+        else:
+            console.print(f"[bold]Fetching {name} for period {period}...[/bold]")
+
+        try:
+            result = fetch_fn(workdir=workdir, period=clamped)
+            results[name] = result
+            console.print(f"[green]{name}: downloaded to datastore[/green]")
+        except FileNotFoundError as exc:
+            # Manual-placement / operator-staged sources can legitimately
+            # be missing on a freshly-bootstrapped project — skip them
+            # rather than fail the whole `fetch all` run so the rest of
+            # the pipeline can proceed. Operators inspect the per-source
+            # skip lines and stage the missing inputs as needed.
+            #   - mwbm-climgrid: CAPTCHA-gated ScienceBase download.
+            #   - daymet: pre-staged zarr root not yet configured.
+            #   - margulis-wus-sr: Oregon-only; legitimately absent for
+            #     non-Oregon projects.
+            manual_skip = {"mwbm-climgrid", "daymet", "margulis-wus-sr"}
+            if name in manual_skip:
+                console.print(
+                    f"[yellow]{name}: skipped (manual download not yet "
+                    f"placed; see docs/sources/{name.replace('-', '_')}.md)[/yellow]"
+                )
+                continue
+            print(f"Error fetching {name}: {exc}", file=sys.stderr)
+            sys.exit(1)
+        except (ValueError, RuntimeError) as exc:
+            print(f"Error fetching {name}: {exc}", file=sys.stderr)
+            sys.exit(1)
+        except Exception as exc:
+            _logger.exception("Unexpected error during %s fetch", name)
+            print(
+                f"Unexpected error fetching {name} ({type(exc).__name__}): {exc}",
+                file=sys.stderr,
+            )
+            sys.exit(1)
+
+    console.print(
+        f"\n[bold green]All {len(results)} sources fetched successfully.[/bold green]"
+    )
+
+
+@fetch_app.command(name="merra2")
+def fetch_merra2_cmd(
+    workdir: Annotated[
+        Path,
+        Parameter(
+            name=["--project-dir"],
+            help="Project created by 'nhf-targets init'.",
+        ),
+    ],
+    period: Annotated[
+        str,
+        Parameter(name=["--period", "-p"], help="Temporal range as 'YYYY/YYYY'."),
+    ],
+):
+    """Download MERRA-2 monthly land surface data (M2TMNXLND).
+
+    Authenticates via earthaccess, searches for granules matching the
+    fabric bounding box, downloads them, and prints the provenance record.
+    """
+    import json as json_mod
+
+    from rich.console import Console
+
+    from nhf_spatial_targets.fetch.merra2 import fetch_merra2
+
+    if not workdir.exists():
+        print(f"Error: Project not found: {workdir}", file=sys.stderr)
+        sys.exit(2)
+
+    console = Console()
+    console.print(f"[bold]Fetching MERRA-2 for period {period}...[/bold]")
+
+    try:
+        result = fetch_merra2(workdir=workdir, period=period)
+    except (ValueError, FileNotFoundError, RuntimeError) as exc:
+        print(f"Error: {exc}", file=sys.stderr)
+        sys.exit(1)
+    except Exception as exc:
+        _logger.exception("Unexpected error during MERRA-2 fetch")
+        print(
+            f"Unexpected error ({type(exc).__name__}): {exc}",
+            file=sys.stderr,
+        )
+        sys.exit(1)
+
+    console.print("[green]MERRA-2: downloaded to datastore[/green]")
+    console.print(json_mod.dumps(result, indent=2))
+
+
+@fetch_app.command(name="nldas-mosaic")
+def fetch_nldas_mosaic_cmd(
+    workdir: Annotated[
+        Path,
+        Parameter(
+            name=["--project-dir"],
+            help="Project created by 'nhf-targets init'.",
+        ),
+    ],
+    period: Annotated[
+        str,
+        Parameter(name=["--period", "-p"], help="Temporal range as 'YYYY/YYYY'."),
+    ],
+):
+    """Download NLDAS-2 MOSAIC soil moisture data."""
+    import json as json_mod
+
+    from rich.console import Console
+
+    from nhf_spatial_targets.fetch.nldas import fetch_nldas_mosaic
+
+    if not workdir.exists():
+        print(f"Error: Project not found: {workdir}", file=sys.stderr)
+        sys.exit(2)
+
+    console = Console()
+    console.print(f"[bold]Fetching NLDAS-2 MOSAIC for period {period}...[/bold]")
+
+    try:
+        result = fetch_nldas_mosaic(workdir=workdir, period=period)
+    except (ValueError, FileNotFoundError, RuntimeError) as exc:
+        print(f"Error: {exc}", file=sys.stderr)
+        sys.exit(1)
+    except Exception as exc:
+        _logger.exception("Unexpected error during NLDAS-2 MOSAIC fetch")
+        print(
+            f"Unexpected error ({type(exc).__name__}): {exc}",
+            file=sys.stderr,
+        )
+        sys.exit(1)
+
+    console.print("[green]NLDAS-2 MOSAIC: downloaded to datastore[/green]")
+    console.print(json_mod.dumps(result, indent=2))
+
+
+@fetch_app.command(name="nldas-noah")
+def fetch_nldas_noah_cmd(
+    workdir: Annotated[
+        Path,
+        Parameter(
+            name=["--project-dir"],
+            help="Project created by 'nhf-targets init'.",
+        ),
+    ],
+    period: Annotated[
+        str,
+        Parameter(name=["--period", "-p"], help="Temporal range as 'YYYY/YYYY'."),
+    ],
+):
+    """Download NLDAS-2 NOAH soil moisture data."""
+    import json as json_mod
+
+    from rich.console import Console
+
+    from nhf_spatial_targets.fetch.nldas import fetch_nldas_noah
+
+    if not workdir.exists():
+        print(f"Error: Project not found: {workdir}", file=sys.stderr)
+        sys.exit(2)
+
+    console = Console()
+    console.print(f"[bold]Fetching NLDAS-2 NOAH for period {period}...[/bold]")
+
+    try:
+        result = fetch_nldas_noah(workdir=workdir, period=period)
+    except (ValueError, FileNotFoundError, RuntimeError) as exc:
+        print(f"Error: {exc}", file=sys.stderr)
+        sys.exit(1)
+    except Exception as exc:
+        _logger.exception("Unexpected error during NLDAS-2 NOAH fetch")
+        print(
+            f"Unexpected error ({type(exc).__name__}): {exc}",
+            file=sys.stderr,
+        )
+        sys.exit(1)
+
+    console.print("[green]NLDAS-2 NOAH: downloaded to datastore[/green]")
+    console.print(json_mod.dumps(result, indent=2))
+
+
+@fetch_app.command(name="ncep-ncar")
+def fetch_ncep_ncar_cmd(
+    workdir: Annotated[
+        Path,
+        Parameter(
+            name=["--project-dir"],
+            help="Project created by 'nhf-targets init'.",
+        ),
+    ],
+    period: Annotated[
+        str,
+        Parameter(name=["--period", "-p"], help="Temporal range as 'YYYY/YYYY'."),
+    ],
+):
+    """Download NCEP/NCAR Reanalysis soil moisture data."""
+    import json as json_mod
+
+    from rich.console import Console
+
+    from nhf_spatial_targets.fetch.ncep_ncar import fetch_ncep_ncar
+
+    if not workdir.exists():
+        print(f"Error: Project not found: {workdir}", file=sys.stderr)
+        sys.exit(2)
+
+    console = Console()
+    console.print(f"[bold]Fetching NCEP/NCAR Reanalysis for period {period}...[/bold]")
+
+    try:
+        result = fetch_ncep_ncar(workdir=workdir, period=period)
+    except (ValueError, FileNotFoundError, RuntimeError) as exc:
+        print(f"Error: {exc}", file=sys.stderr)
+        sys.exit(1)
+    except Exception as exc:
+        _logger.exception("Unexpected error during NCEP/NCAR fetch")
+        print(
+            f"Unexpected error ({type(exc).__name__}): {exc}",
+            file=sys.stderr,
+        )
+        sys.exit(1)
+
+    console.print("[green]NCEP/NCAR: downloaded to datastore[/green]")
+    console.print(json_mod.dumps(result, indent=2))
+
+
+@fetch_app.command(name="mod16a2")
+def fetch_mod16a2_cmd(
+    workdir: Annotated[
+        Path,
+        Parameter(
+            name=["--project-dir"],
+            help="Project created by 'nhf-targets init'.",
+        ),
+    ],
+    period: Annotated[
+        str,
+        Parameter(name=["--period", "-p"], help="Temporal range as 'YYYY/YYYY'."),
+    ],
+    force: Annotated[
+        bool,
+        Parameter(
+            name=["--force"],
+            help=(
+                "Re-fetch every year in --period, ignoring the manifest's "
+                "year-skip. Use after a pipeline change has invalidated the "
+                "existing consolidated NCs (e.g. PR #88's fill-mask fix). The "
+                "manifest entry for mod16a2_v061 is overwritten when the run "
+                "completes."
+            ),
+        ),
+    ] = False,
+):
+    """Download MODIS MOD16A2 v061 AET data (8-day composites, 500m)."""
+    import json as json_mod
+
+    from rich.console import Console
+
+    from nhf_spatial_targets.fetch.modis import fetch_mod16a2
+
+    if not workdir.exists():
+        print(f"Error: Project not found: {workdir}", file=sys.stderr)
+        sys.exit(2)
+
+    console = Console()
+    if force:
+        console.print(
+            f"[bold]Force re-fetching MOD16A2 v061 for period {period}...[/bold]"
+        )
+    else:
+        console.print(f"[bold]Fetching MOD16A2 v061 for period {period}...[/bold]")
+
+    try:
+        result = fetch_mod16a2(workdir=workdir, period=period, force=force)
+    except (ValueError, FileNotFoundError, RuntimeError) as exc:
+        print(f"Error: {exc}", file=sys.stderr)
+        sys.exit(1)
+    except Exception as exc:
+        _logger.exception("Unexpected error during MOD16A2 fetch")
+        print(
+            f"Unexpected error ({type(exc).__name__}): {exc}",
+            file=sys.stderr,
+        )
+        sys.exit(1)
+
+    console.print("[green]MOD16A2: downloaded to datastore[/green]")
+    console.print(json_mod.dumps(result, indent=2))
+
+
+@fetch_app.command(name="mod10c1")
+def fetch_mod10c1_cmd(
+    workdir: Annotated[
+        Path,
+        Parameter(
+            name=["--project-dir"],
+            help="Project created by 'nhf-targets init'.",
+        ),
+    ],
+    period: Annotated[
+        str,
+        Parameter(name=["--period", "-p"], help="Temporal range as 'YYYY/YYYY'."),
+    ],
+    force: Annotated[
+        bool,
+        Parameter(
+            name=["--force"],
+            help=(
+                "Re-fetch every year in --period, ignoring the manifest's "
+                "year-skip. The manifest entry for mod10c1_v061 is "
+                "overwritten when the run completes."
+            ),
+        ),
+    ] = False,
+):
+    """Download MODIS MOD10C1 v061 daily snow cover data (0.05deg CMG)."""
+    import json as json_mod
+
+    from rich.console import Console
+
+    from nhf_spatial_targets.fetch.modis import fetch_mod10c1
+
+    if not workdir.exists():
+        print(f"Error: Project not found: {workdir}", file=sys.stderr)
+        sys.exit(2)
+
+    console = Console()
+    if force:
+        console.print(
+            f"[bold]Force re-fetching MOD10C1 v061 for period {period}...[/bold]"
+        )
+    else:
+        console.print(f"[bold]Fetching MOD10C1 v061 for period {period}...[/bold]")
+
+    try:
+        result = fetch_mod10c1(workdir=workdir, period=period, force=force)
+    except (ValueError, FileNotFoundError, RuntimeError) as exc:
+        print(f"Error: {exc}", file=sys.stderr)
+        sys.exit(1)
+    except Exception as exc:
+        _logger.exception("Unexpected error during MOD10C1 fetch")
+        print(
+            f"Unexpected error ({type(exc).__name__}): {exc}",
+            file=sys.stderr,
+        )
+        sys.exit(1)
+
+    console.print("[green]MOD10C1: downloaded to datastore[/green]")
+    console.print(json_mod.dumps(result, indent=2))
+
+
+@fetch_app.command(name="watergap22d")
+def fetch_watergap22d_cmd(
+    workdir: Annotated[
+        Path,
+        Parameter(
+            name=["--project-dir"],
+            help="Project created by 'nhf-targets init'.",
+        ),
+    ],
+    period: Annotated[
+        str,
+        Parameter(name=["--period", "-p"], help="Temporal range as 'YYYY/YYYY'."),
+    ],
+):
+    """Download WaterGAP 2.2d groundwater recharge from PANGAEA."""
+    import json as json_mod
+
+    from rich.console import Console
+
+    from nhf_spatial_targets.fetch.pangaea import fetch_watergap22d
+
+    if not workdir.exists():
+        print(f"Error: Project not found: {workdir}", file=sys.stderr)
+        sys.exit(2)
+
+    console = Console()
+    console.print(f"[bold]Fetching WaterGAP 2.2d for period {period}...[/bold]")
+
+    try:
+        result = fetch_watergap22d(workdir=workdir, period=period)
+    except (ValueError, FileNotFoundError, RuntimeError) as exc:
+        print(f"Error: {exc}", file=sys.stderr)
+        sys.exit(1)
+    except Exception as exc:
+        _logger.exception("Unexpected error during WaterGAP 2.2d fetch")
+        print(
+            f"Unexpected error ({type(exc).__name__}): {exc}",
+            file=sys.stderr,
+        )
+        sys.exit(1)
+
+    console.print("[green]WaterGAP 2.2d: downloaded to datastore[/green]")
+    console.print(json_mod.dumps(result, indent=2))
+
+
+@fetch_app.command(name="era5-land")
+def fetch_era5_land_cmd(
+    workdir: Annotated[
+        Path,
+        Parameter(
+            name=["--project-dir"],
+            help="Project created by 'nhf-targets init'.",
+        ),
+    ],
+    period: Annotated[
+        str,
+        Parameter(name=["--period", "-p"], help="Temporal range as 'YYYY/YYYY'."),
+    ] = "1979/2024",
+    worker_index: Annotated[
+        int,
+        Parameter(
+            name=["--worker-index"],
+            help="0-based index of this worker within the pool (default 0). "
+            "Set to $SLURM_ARRAY_TASK_ID in array jobs.",
+        ),
+    ] = 0,
+    n_workers: Annotated[
+        int,
+        Parameter(
+            name=["--n-workers"],
+            help="Total number of parallel workers (default 1 = serial). "
+            "Must match the SLURM array size.",
+        ),
+    ] = 1,
+):
+    """Download ERA5-Land hourly fields (ro, sro, ssro, sd) via CDS API and consolidate to daily/monthly NetCDFs.
+
+    The runoff variables (ro/sro/ssro) are accumulated and aggregated as
+    daily/monthly sums; sd (snow depth water equivalent) is instantaneous
+    and aggregated as daily/monthly means.
+    """
+    import json as json_mod
+
+    from rich.console import Console
+
+    from nhf_spatial_targets.fetch.era5_land import fetch_era5_land
+
+    if not workdir.exists():
+        print(f"Error: Project not found: {workdir}", file=sys.stderr)
+        sys.exit(2)
+
+    console = Console()
+    if n_workers > 1:
+        console.print(
+            f"[bold]Fetching ERA5-Land for period {period} "
+            f"(worker {worker_index}/{n_workers})...[/bold]"
+        )
+    else:
+        console.print(f"[bold]Fetching ERA5-Land for period {period}...[/bold]")
+
+    try:
+        result = fetch_era5_land(
+            workdir=workdir,
+            period=period,
+            worker_index=worker_index,
+            n_workers=n_workers,
+        )
+    except (ValueError, FileNotFoundError, RuntimeError) as exc:
+        print(f"Error: {exc}", file=sys.stderr)
+        sys.exit(1)
+    except Exception as exc:
+        _logger.exception("Unexpected error during ERA5-Land fetch")
+        print(
+            f"Unexpected error ({type(exc).__name__}): {exc}",
+            file=sys.stderr,
+        )
+        sys.exit(1)
+
+    console.print("[green]ERA5-Land: downloaded to datastore[/green]")
+    console.print(json_mod.dumps(result, indent=2))
+
+
+@fetch_app.command(name="gldas")
+def fetch_gldas_cmd(
+    workdir: Annotated[
+        Path,
+        Parameter(
+            name=["--project-dir"],
+            help="Project created by 'nhf-targets init'.",
+        ),
+    ],
+    period: Annotated[
+        str,
+        Parameter(name=["--period", "-p"], help="Temporal range as 'YYYY/YYYY'."),
+    ] = "2000/2023",
+):
+    """Download GLDAS-2 monthly runoff data via NASA earthaccess."""
+    import json as json_mod
+
+    from rich.console import Console
+
+    from nhf_spatial_targets.fetch.gldas import fetch_gldas
+
+    if not workdir.exists():
+        print(f"Error: Project not found: {workdir}", file=sys.stderr)
+        sys.exit(2)
+
+    console = Console()
+    console.print(f"[bold]Fetching GLDAS for period {period}...[/bold]")
+
+    try:
+        result = fetch_gldas(workdir=workdir, period=period)
+    except (ValueError, FileNotFoundError, RuntimeError) as exc:
+        print(f"Error: {exc}", file=sys.stderr)
+        sys.exit(1)
+    except Exception as exc:
+        _logger.exception("Unexpected error during GLDAS fetch")
+        print(
+            f"Unexpected error ({type(exc).__name__}): {exc}",
+            file=sys.stderr,
+        )
+        sys.exit(1)
+
+    console.print("[green]GLDAS: downloaded to datastore[/green]")
+    console.print(json_mod.dumps(result, indent=2))
+
+
+@fetch_app.command(name="reitz2017")
+def fetch_reitz2017_cmd(
+    workdir: Annotated[
+        Path,
+        Parameter(
+            name=["--project-dir"],
+            help="Project created by 'nhf-targets init'.",
+        ),
+    ],
+    period: Annotated[
+        str,
+        Parameter(name=["--period", "-p"], help="Temporal range as 'YYYY/YYYY'."),
+    ],
+):
+    """Download Reitz 2017 annual recharge estimates from USGS ScienceBase."""
+    import json as json_mod
+
+    from rich.console import Console
+
+    from nhf_spatial_targets.fetch.reitz2017 import fetch_reitz2017
+
+    if not workdir.exists():
+        print(f"Error: Project not found: {workdir}", file=sys.stderr)
+        sys.exit(2)
+
+    console = Console()
+    console.print(f"[bold]Fetching Reitz 2017 recharge for period {period}...[/bold]")
+
+    try:
+        result = fetch_reitz2017(workdir=workdir, period=period)
+    except (ValueError, FileNotFoundError, RuntimeError) as exc:
+        print(f"Error: {exc}", file=sys.stderr)
+        sys.exit(1)
+    except Exception as exc:
+        _logger.exception("Unexpected error during Reitz 2017 fetch")
+        print(
+            f"Unexpected error ({type(exc).__name__}): {exc}",
+            file=sys.stderr,
+        )
+        sys.exit(1)
+
+    console.print("[green]Reitz 2017: downloaded to datastore[/green]")
+    console.print(json_mod.dumps(result, indent=2))
+
+
+@fetch_app.command(name="mwbm-climgrid")
+def fetch_mwbm_climgrid_cmd(
+    workdir: Annotated[
+        Path,
+        Parameter(
+            name=["--project-dir"],
+            help="Project created by 'nhf-targets init'.",
+        ),
+    ],
+    period: Annotated[
+        str,
+        Parameter(name=["--period", "-p"], help="Temporal range as 'YYYY/YYYY'."),
+    ],
+):
+    """Register a manually-placed USGS MWBM (ClimGrid-forced) NetCDF.
+
+    The ScienceBase distribution is gated by a CAPTCHA, so the ~7.5 GB
+    ClimGrid_WBM.nc cannot be retrieved automatically. Download it via
+    a browser and place it at <datastore>/mwbm_climgrid/ClimGrid_WBM.nc
+    before invoking this command — see docs/sources/mwbm_climgrid.md
+    for the procedure. This command then fingerprints the file and
+    writes its provenance to manifest.json.
+    """
+    import json as json_mod
+
+    from rich.console import Console
+
+    from nhf_spatial_targets.fetch.mwbm_climgrid import fetch_mwbm_climgrid
+
+    if not workdir.exists():
+        print(f"Error: Project not found: {workdir}", file=sys.stderr)
+        sys.exit(2)
+
+    console = Console()
+    console.print(f"[bold]Registering MWBM ClimGrid (period {period})...[/bold]")
+
+    try:
+        result = fetch_mwbm_climgrid(workdir=workdir, period=period)
+    except (ValueError, FileNotFoundError, RuntimeError) as exc:
+        print(f"Error: {exc}", file=sys.stderr)
+        sys.exit(1)
+    except Exception as exc:
+        _logger.exception("Unexpected error during MWBM ClimGrid registration")
+        print(
+            f"Unexpected error ({type(exc).__name__}): {exc}",
+            file=sys.stderr,
+        )
+        sys.exit(1)
+
+    console.print("[green]MWBM ClimGrid: registered in manifest[/green]")
+    console.print(json_mod.dumps(result, indent=2))
+
+
+# ---------------------------------------------------------------------------
+# SWE category (issue #99) — fetch scaffolding
+# Aggregate + target wiring deferred to per-source follow-up issues.
+# ---------------------------------------------------------------------------
+
+
+@fetch_app.command(name="daymet")
+def fetch_daymet_cmd(
+    workdir: Annotated[
+        Path,
+        Parameter(
+            name=["--project-dir"],
+            help="Project created by 'nhf-targets init'.",
+        ),
+    ],
+    period: Annotated[
+        str,
+        Parameter(name=["--period", "-p"], help="Temporal range as 'YYYY/YYYY'."),
+    ] = "1980/2024",
+    source_path: Annotated[
+        Path | None,
+        Parameter(
+            name=["--source-path"],
+            help="Directory containing daymet_{na,hi,pr}.zarr. Overrides "
+            "'daymet_root' from config.yml.",
+        ),
+    ] = None,
+    region: Annotated[
+        str,
+        Parameter(
+            name=["--region"],
+            help="One of 'na' | 'hi' | 'pr' | 'all' (default: all).",
+        ),
+    ] = "all",
+):
+    """Register operator-staged Daymet V4 R1 regional zarr stores.
+
+    Daymet zarrs are pre-staged on a shared filesystem; this command
+    fingerprints the structural metadata of each region and records a
+    per-region manifest entry. No downloading or copying. See
+    docs/sources/daymet.md for the staging procedure.
+    """
+    import json as json_mod
+
+    from rich.console import Console
+
+    from nhf_spatial_targets.fetch.daymet import fetch_daymet
+
+    if not workdir.exists():
+        print(f"Error: Project not found: {workdir}", file=sys.stderr)
+        sys.exit(2)
+
+    console = Console()
+    console.print(
+        f"[bold]Registering Daymet (region={region}, period {period})...[/bold]"
+    )
+
+    try:
+        result = fetch_daymet(
+            workdir=workdir,
+            period=period,
+            source_path=source_path,
+            region=region,
+        )
+    except (ValueError, FileNotFoundError, RuntimeError) as exc:
+        print(f"Error: {exc}", file=sys.stderr)
+        sys.exit(1)
+    except Exception as exc:
+        _logger.exception("Unexpected error during Daymet registration")
+        print(
+            f"Unexpected error ({type(exc).__name__}): {exc}",
+            file=sys.stderr,
+        )
+        sys.exit(1)
+
+    console.print("[green]Daymet: registered in manifest[/green]")
+    console.print(json_mod.dumps(result, indent=2))
+
+
+@fetch_app.command(name="snodas")
+def fetch_snodas_cmd(
+    workdir: Annotated[
+        Path,
+        Parameter(
+            name=["--project-dir"],
+            help="Project created by 'nhf-targets init'.",
+        ),
+    ],
+    period: Annotated[
+        str,
+        Parameter(name=["--period", "-p"], help="Temporal range as 'YYYY/YYYY'."),
+    ] = "2003/2024",
+    worker_index: Annotated[
+        int,
+        Parameter(
+            name=["--worker-index"],
+            help="0-based index of this worker within the pool (default 0).",
+        ),
+    ] = 0,
+    n_workers: Annotated[
+        int,
+        Parameter(
+            name=["--n-workers"],
+            help="Total number of parallel workers (default 1 = serial).",
+        ),
+    ] = 1,
+):
+    """Download SNODAS daily SWE granules from NSIDC (G02158) via earthaccess.
+
+    Fetch-only: raw tar/.Hdr bundles land in <datastore>/snodas/raw/<year>/.
+    Decoding into CF NetCDFs is deferred to the SNODAS aggregate follow-up.
+    """
+    import json as json_mod
+
+    from rich.console import Console
+
+    from nhf_spatial_targets.fetch.snodas import fetch_snodas
+
+    if not workdir.exists():
+        print(f"Error: Project not found: {workdir}", file=sys.stderr)
+        sys.exit(2)
+
+    console = Console()
+    if n_workers > 1:
+        console.print(
+            f"[bold]Fetching SNODAS for period {period} "
+            f"(worker {worker_index}/{n_workers})...[/bold]"
+        )
+    else:
+        console.print(f"[bold]Fetching SNODAS for period {period}...[/bold]")
+
+    try:
+        result = fetch_snodas(
+            workdir=workdir,
+            period=period,
+            worker_index=worker_index,
+            n_workers=n_workers,
+        )
+    except (ValueError, FileNotFoundError, RuntimeError) as exc:
+        print(f"Error: {exc}", file=sys.stderr)
+        sys.exit(1)
+    except Exception as exc:
+        _logger.exception("Unexpected error during SNODAS fetch")
+        print(
+            f"Unexpected error ({type(exc).__name__}): {exc}",
+            file=sys.stderr,
+        )
+        sys.exit(1)
+
+    console.print("[green]SNODAS: downloaded to datastore[/green]")
+    console.print(json_mod.dumps(result, indent=2))
+
+
+@fetch_app.command(name="margulis-wus-sr")
+def fetch_margulis_wus_sr_cmd(
+    workdir: Annotated[
+        Path,
+        Parameter(
+            name=["--project-dir"],
+            help="Project created by 'nhf-targets init'.",
+        ),
+    ],
+    period: Annotated[
+        str,
+        Parameter(name=["--period", "-p"], help="Temporal range as 'YYYY/YYYY'."),
+    ] = "1985/2021",
+):
+    """Download Margulis Western US Snow Reanalysis (NSIDC-0719) via earthaccess.
+
+    Fabric-scoped to Oregon only (catalog `fabric_scope`); the scope is
+    recorded in manifest.json but not enforced at fetch time. Fetch-only:
+    consolidation is deferred to the Margulis aggregate follow-up.
+    """
+    import json as json_mod
+
+    from rich.console import Console
+
+    from nhf_spatial_targets.fetch.margulis_wus_sr import fetch_margulis_wus_sr
+
+    if not workdir.exists():
+        print(f"Error: Project not found: {workdir}", file=sys.stderr)
+        sys.exit(2)
+
+    console = Console()
+    console.print(f"[bold]Fetching Margulis WUS-SR for period {period}...[/bold]")
+
+    try:
+        result = fetch_margulis_wus_sr(workdir=workdir, period=period)
+    except (ValueError, FileNotFoundError, RuntimeError) as exc:
+        print(f"Error: {exc}", file=sys.stderr)
+        sys.exit(1)
+    except Exception as exc:
+        _logger.exception("Unexpected error during Margulis WUS-SR fetch")
+        print(
+            f"Unexpected error ({type(exc).__name__}): {exc}",
+            file=sys.stderr,
+        )
+        sys.exit(1)
+
+    console.print("[green]Margulis WUS-SR: downloaded to datastore[/green]")
+    console.print(json_mod.dumps(result, indent=2))
